@@ -14,9 +14,9 @@ class QuadHexagon:
         """Time to open a `Grid`"""
         ux.open_grid(grid_path)
 
-    def mem_open_grid(self):
-        """Memory Occupied by a `Grid`"""
-        return ux.open_grid(grid_path)
+    # def mem_open_grid(self):
+    #     """Memory Occupied by a `Grid`"""
+    #     return ux.open_grid(grid_path)
 
     def peakmem_open_grid(self):
         """Peak memory usage of a `Grid`"""
@@ -27,9 +27,9 @@ class QuadHexagon:
         """Time to open a `UxDataset`"""
         ux.open_dataset(grid_path, data_path)
 
-    def mem_open_dataset(self):
-        """Memory occupied by a `UxDataset`"""
-        return ux.open_dataset(grid_path, data_path)
+    # def mem_open_dataset(self):
+    #     """Memory occupied by a `UxDataset`"""
+    #     return ux.open_dataset(grid_path, data_path)
 
     def peakmem_open_dataset(self):
         """Peak memory usage of a `UxDataset`"""
