@@ -52,29 +52,12 @@ extensions = [
     "sphinx.ext.mathjax",
     "sphinx_autosummary_accessors",
     "myst_nb",
-    "sphinx_design",
     "nbsphinx",
+    "sphinx_design",
     "IPython.sphinxext.ipython_directive",
     "IPython.sphinxext.ipython_console_highlighting",
+    "sphinx_remove_toctrees",
 ]
-
-mathjax_config = {
-    "tex2jax": {
-        "inlineMath": [["$", "$"], ["\\(", "\\)"]],
-        "displayMath": [["$$", "$$"], ["\\[", "\\]"]],
-        "processEscapes": True,
-    },
-    "HTML-CSS": {
-        "linebreaks": {"automatic": True, "width": "container"},
-        "fonts": {"availableFonts": ["TeX"], "preferredFont": "TeX", "webFont": "TeX"},
-    },
-    "CommonHTML": {"linebreaks": {"automatic": True, "width": "container"}},
-    "SVG": {"linebreaks": {"automatic": True, "width": "container"}},
-}
-
-mathjax_path = (
-    "https://cdn.jsdelivr.net/npm/mathjax@2/MathJax.js?config=TeX-AMS-MML_HTMLorMML"
-)
 
 extlinks = {
     "issue": ("https://github.com/uxarray/uxarray/issues/%s", "GH"),
@@ -96,6 +79,8 @@ intersphinx_mapping = {
     "xarray": ("http://xarray.pydata.org/en/stable/", None),
 }
 
+remove_from_toctrees = ["generated/*"]
+
 napoleon_use_admonition_for_examples = True
 napoleon_include_special_with_doc = True
 
@@ -113,6 +98,13 @@ source_suffix = {
     ".myst": "myst-nb",
 }
 
+# MyST configuration
+myst_enable_extensions = [
+    "amsmath",  # for extended math environments
+    "dollarmath",  # for $ ... $ syntax
+]
+
+
 # The master toctree document.
 master_doc = "index"
 
@@ -120,7 +112,7 @@ master_doc = "index"
 project = "uxarray"
 
 current_year = datetime.datetime.now().year
-copyright = "{}, Uxarray".format(current_year)
+copyright = "{}, UXarray".format(current_year)
 author = "UXARRAY"
 
 # The language for content autogenerated by Sphinx. Refer to documentation
@@ -160,19 +152,22 @@ html_context = {
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
 # documentation.
-html_theme_options = dict(
-    # analytics_id=''  this is configured in rtfd.io
-    # canonical_url="",
-    repository_url="https://github.com/uxarray/uxarray",
-    repository_branch="main",
-    path_to_docs="docs",
-    use_edit_page_button=True,
-    use_repository_button=True,
-    use_issues_button=True,
-    home_page_in_toc=False,
-    navbar_footer_text="",
-    # extra_footer="""<p></p>""",
-)
+html_theme_options = {
+    "repository_url": "https://github.com/uxarray/uxarray",
+    "repository_branch": "main",
+    "path_to_docs": "docs",
+    "use_edit_page_button": True,
+    "use_repository_button": True,
+    "use_issues_button": True,
+    "home_page_in_toc": False,
+    "navbar_footer_text": "",
+    # "extra_footer": "<p></p>",
+    "logo": {
+        "image_light": "_static/images/logos/uxarray_logo_h_dark.svg",
+        "image_dark": "_static/images/logos/uxarray_logo_h_light.svg",
+    },
+}
+
 
 # The name for this set of Sphinx documents.
 # "<project> v<release> documentation" by default.
@@ -180,58 +175,67 @@ html_theme_options = dict(
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
-html_logo = "_static/images/logos/uxarray_temp_logo.png"
+html_logo = "_static/images/logos/uxarray_logo_h_dark.svg"
 
 # The name of an image file (within the static path) to use as favicon of the
 # docs.  This file should be a Windows icon file (.ico) being 16x16 or 32x32
 # pixels large.
-html_favicon = "_static/images/logos/uxarray_temp_logo.png"
+html_favicon = "_static/images/logos/uxarray_logo_h_dark.svg"
+
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
-html_css_files = ["style.css"]
+# html_css_files = ["style.css"]
 
 # Output file base name for HTML help builder.
 htmlhelp_basename = "uxarraydoc"
 
 autodoc_typehints = "none"
 
+# notebooks to exclude from run-time execution, will use rendered version
+nb_execution_excludepatterns = [
+    "3_75km_mpas.ipynb",
+    "e3sm-calc-workflow.ipynb",
+    "e3sm-load-viz.ipynb",
+    "parallel-load-ux-with-dask.ipynb",
+]
+
 
 # custom scripts for making a gallery of examples notebooks
 def update_gallery(app: Sphinx):
-    """Update the gallery of examples notebooks."""
+    """Update the gallery page."""
 
-    LOGGER.info("creating gallery...")
+    LOGGER.info("Updating gallery page...")
 
-    notebooks = yaml.safe_load(pathlib.Path(app.srcdir, "gallery.yml").read_bytes())
+    gallery = yaml.safe_load(pathlib.Path(app.srcdir, "gallery.yml").read_bytes())
 
-    items = [
-        f"""
+    for key in gallery:
+        items = [
+            f"""
          .. grid-item-card::
             :text-align: center
-            :link: {item['path']}
+            :link: {item["path"]}
 
-            .. image:: {item['thumbnail']}
-                :alt: {item['title']}
+            .. image:: {item["thumbnail"]}
+                :alt: {item["title"]}
             +++
-            {item['title']}
+            {item["title"]}
             """
-        for item in notebooks
-    ]
+            for item in gallery[key]
+        ]
 
-    items_md = indent(dedent("\n".join(items)), prefix="    ")
-    markdown = f"""
+        items_md = indent(dedent("\n".join(items)), prefix="    ")
+        markdown = f"""
 .. grid:: 1 2 3 3
     :gutter: 2
 
     {items_md}
     """
-
-    pathlib.Path(app.srcdir, "notebook-examples.txt").write_text(markdown)
-
-    LOGGER.info("gallery created")
+        pathlib.Path(app.srcdir, f"{key}-gallery.txt").write_text(markdown)
+        LOGGER.info(f"{key} gallery page updated.")
+    LOGGER.info("Gallery page updated.")
 
 
 # Allow for changes to be made to the css in the theme_overrides file
