@@ -6,7 +6,8 @@ NNODES_outCSne8 = 386
 NNODES_outCSne30 = 5402
 NNODES_outRLL1deg = 64442
 DATAVARS_outCSne30 = 4
-TRI_AREA = 1.047
+TRI_AREA = 0.02216612469199045
+CORRECTED_TRI_AREA = 0.02244844510268421
 # 4*Pi is 12.56
 MESH30_AREA = 12.566
 PSI_INTG = 12.566
