@@ -1,141 +1,106 @@
 import os
-
-from unittest import TestCase
 from pathlib import Path
-
 import numpy as np
-
 import uxarray as ux
-
 from uxarray.grid.geometry import _build_polygon_shells, _build_corrected_polygon_shells
-
-from uxarray.core.dataset import UxDataset
+from uxarray.core.dataset import UxDataset, UxDataArray
+import pytest
 
 current_path = Path(os.path.dirname(os.path.realpath(__file__)))
 
 gridfile_ne30 = current_path / "meshfiles" / "ugrid" / "outCSne30" / "outCSne30.ug"
 dsfile_var2_ne30 = current_path / "meshfiles" / "ugrid" / "outCSne30" / "outCSne30_vortex.nc"
 
-dsfiles_mf_ne30 = str(
-    current_path) + "/meshfiles/ugrid/outCSne30/outCSne30_*.nc"
-
 gridfile_geoflow = current_path / "meshfiles" / "ugrid" / "geoflow-small" / "grid.nc"
 dsfile_v1_geoflow = current_path / "meshfiles" / "ugrid" / "geoflow-small" / "v1.nc"
 
+def test_to_dataset():
+    """Tests the conversion of UxDataArrays to a UXDataset."""
+    uxds = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
+    uxds_converted = uxds['psi'].to_dataset()
 
-class TestDataArray(TestCase):
+    assert isinstance(uxds_converted, UxDataset)
+    assert uxds_converted.uxgrid == uxds.uxgrid
 
-    def test_to_dataset(self):
-        """Tests the conversion of UxDataArrays to a UXDataset."""
-        uxds = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
-        uxds_converted = uxds['psi'].to_dataset()
+def test_get_dual():
+    """Tests the creation of the dual mesh on a data array."""
+    uxds = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
+    dual = uxds['psi'].get_dual()
 
-        assert isinstance(uxds_converted, UxDataset)
-        assert uxds_converted.uxgrid == uxds.uxgrid
+    assert isinstance(dual, UxDataArray)
+    assert dual._node_centered()
 
+def test_to_geodataframe():
+    """Tests the conversion to ``GeoDataFrame``"""
+    # GeoFlow
+    uxds_geoflow = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
 
-class TestGeometryConversions(TestCase):
+    # v1 is mapped to nodes, should raise a value error
+    with pytest.raises(ValueError):
+        uxds_geoflow['v1'].to_geodataframe()
 
-    def test_to_geodataframe(self):
-        """Tests the conversion to ``GeoDataFrame``"""
-        ### geoflow
-        uxds_geoflow = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    # grid conversion
+    gdf_geoflow_grid = uxds_geoflow.uxgrid.to_geodataframe(periodic_elements='split')
 
-        # v1 is mapped to nodes, should raise a value error
-        with self.assertRaises(ValueError):
-            uxds_geoflow['v1'].to_geodataframe()
+    # number of elements
+    assert gdf_geoflow_grid.shape == (uxds_geoflow.uxgrid.n_face, 1)
 
-        # grid conversion
-        gdf_geoflow_grid = uxds_geoflow.uxgrid.to_geodataframe()
+    # NE30
+    uxds_ne30 = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
 
-        # number of elements
-        assert gdf_geoflow_grid.shape == (uxds_geoflow.uxgrid.n_face, 1)
+    gdf_geoflow_data = uxds_ne30['psi'].to_geodataframe(periodic_elements='split')
 
-        ### n30
-        uxds_ne30 = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
+    assert gdf_geoflow_data.shape == (uxds_ne30.uxgrid.n_face, 2)
 
-        gdf_geoflow_data = uxds_ne30['psi'].to_geodataframe()
+def test_to_polycollection():
+    """Tests the conversion to ``PolyCollection``"""
+    # GeoFlow
+    uxds_geoflow = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
 
-        assert gdf_geoflow_data.shape == (uxds_ne30.uxgrid.n_face, 2)
+    # v1 is mapped to nodes, should raise a value error
+    with pytest.raises(ValueError):
+        uxds_geoflow['v1'].to_polycollection()
 
-    def test_to_polycollection(self):
-        """Tests the conversion to ``PolyCollection``"""
-        ### geoflow
-        uxds_geoflow = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    # grid conversion
+    pc_geoflow_grid = uxds_geoflow.uxgrid.to_polycollection(periodic_elements='split')
 
-        # v1 is mapped to nodes, should raise a value error
-        with self.assertRaises(ValueError):
-            uxds_geoflow['v1'].to_polycollection()
+    polygon_shells = _build_polygon_shells(
+        uxds_geoflow.uxgrid.node_lon.values,
+        uxds_geoflow.uxgrid.node_lat.values,
+        uxds_geoflow.uxgrid.face_node_connectivity.values,
+        uxds_geoflow.uxgrid.n_face, uxds_geoflow.uxgrid.n_max_face_nodes,
+        uxds_geoflow.uxgrid.n_nodes_per_face.values)
 
-        # grid conversion
-        pc_geoflow_grid, _ = uxds_geoflow.uxgrid.to_polycollection()
+    corrected_polygon_shells, _ = _build_corrected_polygon_shells(polygon_shells)
 
-        polygon_shells = _build_polygon_shells(
-            uxds_geoflow.uxgrid.node_lon.values,
-            uxds_geoflow.uxgrid.node_lat.values,
-            uxds_geoflow.uxgrid.face_node_connectivity.values,
-            uxds_geoflow.uxgrid.n_face, uxds_geoflow.uxgrid.n_max_face_nodes,
-            uxds_geoflow.uxgrid.n_nodes_per_face.values)
+    # number of elements
+    assert len(pc_geoflow_grid._paths) == len(corrected_polygon_shells)
 
-        corrected_polygon_shells, _ = _build_corrected_polygon_shells(
-            polygon_shells)
+    # NE30
+    uxds_ne30 = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
 
-        # number of elements
-        assert len(pc_geoflow_grid._paths) == len(corrected_polygon_shells)
+    polygon_shells = _build_polygon_shells(
+        uxds_ne30.uxgrid.node_lon.values, uxds_ne30.uxgrid.node_lat.values,
+        uxds_ne30.uxgrid.face_node_connectivity.values,
+        uxds_ne30.uxgrid.n_face, uxds_ne30.uxgrid.n_max_face_nodes,
+        uxds_ne30.uxgrid.n_nodes_per_face.values)
 
-        # ### n30
-        uxds_ne30 = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
+    corrected_polygon_shells, _ = _build_corrected_polygon_shells(polygon_shells)
 
-        polygon_shells = _build_polygon_shells(
-            uxds_ne30.uxgrid.node_lon.values, uxds_ne30.uxgrid.node_lat.values,
-            uxds_ne30.uxgrid.face_node_connectivity.values,
-            uxds_ne30.uxgrid.n_face, uxds_ne30.uxgrid.n_max_face_nodes,
-            uxds_ne30.uxgrid.n_nodes_per_face.values)
+    pc_geoflow_data = uxds_ne30['psi'].to_polycollection(periodic_elements='split')
 
-        corrected_polygon_shells, _ = _build_corrected_polygon_shells(
-            polygon_shells)
+    assert len(pc_geoflow_data._paths) == len(corrected_polygon_shells)
 
-        pc_geoflow_data, _ = uxds_ne30['psi'].to_polycollection()
+def test_geodataframe_caching():
+    uxds = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
 
-        assert len(pc_geoflow_data._paths) == len(corrected_polygon_shells)
+    gdf_start = uxds['psi'].to_geodataframe()
+    gdf_next = uxds['psi'].to_geodataframe()
 
-    def test_geodataframe_caching(self):
-        uxds = ux.open_dataset(gridfile_ne30, dsfile_var2_ne30)
+    # with caching, they point to the same area in memory
+    assert gdf_start is gdf_next
 
-        gdf_start = uxds['psi'].to_geodataframe()
+    gdf_end = uxds['psi'].to_geodataframe(override=True)
 
-        gdf_next = uxds['psi'].to_geodataframe()
-
-        # with caching, they point to the same area in memory
-        assert gdf_start is gdf_next
-
-        gdf_end = uxds['psi'].to_geodataframe(override=True)
-
-        # override will recompute the grid
-        assert gdf_start is not gdf_end
-
-    def test_nodal_average(self):
-
-        # test on a node-centered dataset
-        uxds = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
-
-        v1_nodal_average = uxds['v1'].nodal_average()
-
-        # final dimension should match number of faces
-        self.assertEqual(v1_nodal_average.shape[-1], uxds.uxgrid.n_face)
-
-        # all other dimensions should remain unchanged
-        self.assertEqual(uxds['v1'].shape[0:-1], v1_nodal_average.shape[0:-1])
-
-        # test on a sample mesh with 4 verts
-        verts = [[[-170, 40], [180, 30], [165, 25], [-170, 20]]]
-        data = [1, 2, 3, 4]
-
-        uxgrid = ux.open_grid(verts, latlon=True)
-
-        uxda = ux.UxDataArray(uxgrid=uxgrid, data=data, dims=('n_node'))
-
-        uxda_nodal_average = uxda.nodal_average()
-
-        # resulting data should be the mean of the corner nodes of the single face
-        self.assertEqual(uxda_nodal_average, np.mean(data))
+    # override will recompute the grid
+    assert gdf_start is not gdf_end
