@@ -1,17 +1,12 @@
 import sys
-from unittest import TestCase
-
 import xarray as xr
 
-# Import from directory structure if coverage test, or from installed
-# packages otherwise
+# Import from directory structure if coverage test, or from installed packages otherwise
 if "--cov" in str(sys.argv):
     import uxarray
 else:
     import uxarray
 
-
-class test_placeholder(TestCase):
-
-    def test_placeholder(self):
-        pass
+def test_placeholder():
+    """Placeholder test that currently does nothing."""
+    pass
