@@ -1,114 +1,99 @@
 import os
 import uxarray as ux
-
-from unittest import TestCase
+import holoviews as hv
+import pytest
 from pathlib import Path
 
 current_path = Path(os.path.dirname(os.path.realpath(__file__)))
 
 gridfile_geoflow = current_path / "meshfiles" / "ugrid" / "geoflow-small" / "grid.nc"
 datafile_geoflow = current_path / "meshfiles" / "ugrid" / "geoflow-small" / "v1.nc"
-
 gridfile_mpas = current_path / "meshfiles" / "mpas" / "QU" / "oQU480.231010.nc"
-
 gridfile_ne30 = current_path / "meshfiles" / "ugrid" / "outCSne30" / "outCSne30.ug"
-
 datafile_ne30 = current_path / "meshfiles" / "ugrid" / "outCSne30" / "outCSne30_vortex.nc"
 
 grid_files = [gridfile_geoflow, gridfile_mpas]
+grid_plot_routines = ['points', 'nodes', 'node_coords', '']
 
+def test_topology():
+    """Tests execution on Grid elements."""
+    uxgrid = ux.open_grid(gridfile_mpas)
 
-class TestPlot(TestCase):
+    for backend in ['matplotlib', 'bokeh']:
+        uxgrid.plot(backend=backend)
+        uxgrid.plot.mesh(backend=backend)
+        uxgrid.plot.edges(backend=backend)
+        uxgrid.plot.nodes(backend=backend)
+        uxgrid.plot.node_coords(backend=backend)
+        uxgrid.plot.corner_nodes(backend=backend)
+        uxgrid.plot.face_centers(backend=backend)
+        uxgrid.plot.face_coords(backend=backend)
+        uxgrid.plot.edge_centers(backend=backend)
+        uxgrid.plot.edge_coords(backend=backend)
 
-    def test_topology(self):
-        """Tests execution on Grid elements."""
-        uxgrid = ux.open_grid(gridfile_mpas)
+def test_face_centered_data():
+    """Tests execution of plotting methods on face-centered data."""
+    uxds = ux.open_dataset(gridfile_mpas, gridfile_mpas)
 
-        for backend in ['matplotlib', 'bokeh']:
+    for backend in ['matplotlib', 'bokeh']:
+        assert isinstance(uxds['bottomDepth'].plot(backend=backend, dynamic=True), hv.DynamicMap)
+        assert isinstance(uxds['bottomDepth'].plot.polygons(backend=backend, dynamic=True), hv.DynamicMap)
+        assert isinstance(uxds['bottomDepth'].plot.points(backend=backend), hv.Points)
 
-            uxgrid.plot(backend=backend)
+def test_face_centered_remapped_dim():
+    """Tests execution of plotting method on a data variable whose dimension needed to be re-mapped."""
+    uxds = ux.open_dataset(gridfile_ne30, datafile_ne30)
 
-            uxgrid.plot.mesh(backend=backend)
+    for backend in ['matplotlib', 'bokeh']:
+        assert isinstance(uxds['psi'].plot(backend=backend, dynamic=True), hv.DynamicMap)
+        assert isinstance(uxds['psi'].plot.polygons(backend=backend, dynamic=True), hv.DynamicMap)
+        assert isinstance(uxds['psi'].plot.points(backend=backend), hv.Points)
 
-            uxgrid.plot.edges(backend=backend)
+def test_node_centered_data():
+    """Tests execution of plotting methods on node-centered data."""
+    uxds = ux.open_dataset(gridfile_geoflow, datafile_geoflow)
 
-            uxgrid.plot.nodes(backend=backend)
+    for backend in ['matplotlib', 'bokeh']:
+        assert isinstance(uxds['v1'][0][0].plot(backend=backend), hv.Points)
+        assert isinstance(uxds['v1'][0][0].plot.points(backend=backend), hv.Points)
+        assert isinstance(uxds['v1'][0][0].topological_mean(destination='face').plot.polygons(backend=backend, dynamic=True), hv.DynamicMap)
 
-            uxgrid.plot.face_centers(backend=backend)
+def test_clabel():
+    """Tests the execution of passing in a custom clabel."""
+    uxds = ux.open_dataset(gridfile_geoflow, datafile_geoflow)
 
-            if uxgrid.edge_lon is not None:
-                uxgrid.plot.edge_centers(backend=backend)
+    raster_no_clabel = uxds['v1'][0][0].plot.rasterize(method='point')
+    raster_with_clabel = uxds['v1'][0][0].plot.rasterize(method='point', clabel='Foo')
 
-    def test_face_centered_data(self):
-        """Tests execution of plotting methods on face-centered data."""
+def test_engine():
+    """Tests different plotting engines."""
+    uxds = ux.open_dataset(gridfile_mpas, gridfile_mpas)
+    _plot_sp = uxds['bottomDepth'].plot.polygons(rasterize=True, dynamic=True, engine='spatialpandas')
+    _plot_gp = uxds['bottomDepth'].plot.polygons(rasterize=True, dynamic=True, engine='geopandas')
 
-        uxds = ux.open_dataset(gridfile_mpas, gridfile_mpas)
+    assert isinstance(_plot_sp, hv.DynamicMap)
+    assert isinstance(_plot_gp, hv.DynamicMap)
 
-        for backend in ['matplotlib', 'bokeh']:
+def test_dataset_methods():
+    """Tests whether a Xarray DataArray method can be called through the UxDataArray plotting accessor."""
+    uxds = ux.open_dataset(gridfile_geoflow, datafile_geoflow)
 
-            uxds['bottomDepth'].plot(backend=backend)
+    # plot.hist() is an xarray method
+    assert hasattr(uxds['v1'].plot, 'hist')
 
-            uxds['bottomDepth'].plot.polygons(backend=backend)
+def test_dataarray_methods():
+    """Tests whether a Xarray Dataset method can be called through the UxDataset plotting accessor."""
+    uxds = ux.open_dataset(gridfile_geoflow, datafile_geoflow)
 
-            uxds['bottomDepth'].plot.points(backend=backend)
+    # plot.scatter() is an xarray method
+    assert hasattr(uxds.plot, 'scatter')
 
-            uxds['bottomDepth'].plot.rasterize(method='polygon',
-                                               backend=backend)
+def test_line():
+    uxds = ux.open_dataset(gridfile_mpas, gridfile_mpas)
+    _plot_line = uxds['bottomDepth'].zonal_average().plot.line()
+    assert isinstance(_plot_line, hv.Curve)
 
-    def test_face_centered_remapped_dim(self):
-        """Tests execution of plotting method on a data variable whose
-        dimension needed to be re-mapped."""
-        uxds = ux.open_dataset(gridfile_ne30, datafile_ne30)
-
-        for backend in ['matplotlib', 'bokeh']:
-
-            uxds['psi'].plot(backend=backend)
-
-            uxds['psi'].plot.polygons(backend=backend)
-
-            uxds['psi'].plot.points(backend=backend)
-
-            uxds['psi'].plot.rasterize(method='polygon', backend=backend)
-
-    def test_node_centered_data(self):
-        """Tests execution of plotting methods on node-centered data."""
-
-        uxds = ux.open_dataset(gridfile_geoflow, datafile_geoflow)
-
-        for backend in ['matplotlib', 'bokeh']:
-            uxds['v1'][0][0].plot(backend=backend)
-
-            uxds['v1'][0][0].plot.points(backend=backend)
-
-            uxds['v1'][0][0].nodal_average().plot.polygons(backend=backend)
-
-
-    def test_clabel(self):
-        """Tests the execution of passing in a custom clabel."""
-
-        uxds = ux.open_dataset(gridfile_geoflow, datafile_geoflow)
-
-        raster_no_clabel = uxds['v1'][0][0].plot.rasterize(method='point')
-
-        raster_with_clabel = uxds['v1'][0][0].plot.rasterize(method='point', clabel='Foo')
-
-
-
-class TestXarrayMethods(TestCase):
-
-    def test_dataset(self):
-        """Tests whether a Xarray DataArray method can be called through the
-        UxDataArray plotting accessor."""
-        uxds = ux.open_dataset(gridfile_geoflow, datafile_geoflow)
-
-        # plot.hist() is an xarray method
-        assert hasattr(uxds['v1'].plot, 'hist')
-
-
-    def test_dataarray(self):
-        """Tests whether a Xarray Dataset method can be called through the
-        UxDataset plotting accessor."""
-        uxds = ux.open_dataset(gridfile_geoflow, datafile_geoflow)
-
-        # plot.scatter() is an xarray method
-        assert hasattr(uxds.plot, 'scatter')
+def test_scatter():
+    uxds = ux.open_dataset(gridfile_mpas, gridfile_mpas)
+    _plot_line = uxds['bottomDepth'].zonal_average().plot.scatter()
+    assert isinstance(_plot_line, hv.Scatter)
