@@ -1,14 +1,12 @@
 import os
 import numpy as np
-
-from unittest import TestCase
+import numpy.testing as nt
+import pytest
 from pathlib import Path
 
 import uxarray as ux
-
 from uxarray.core.dataset import UxDataset
 from uxarray.core.dataarray import UxDataArray
-
 from uxarray.remap.inverse_distance_weighted import _inverse_distance_weighted_remap
 from uxarray.remap.nearest_neighbor import _nearest_neighbor
 
@@ -23,270 +21,246 @@ dsfile_v3_geoflow = current_path / "meshfiles" / "ugrid" / "geoflow-small" / "v3
 mpasfile_QU = current_path / "meshfiles" / "mpas" / "QU" / "mesh.QU.1920km.151026.nc"
 
 
-class TestNearestNeighborRemap(TestCase):
-    """Tests for nearest neighbor remapping."""
-
-    def test_remap_to_same_grid_corner_nodes(self):
-        """Test remapping to the same dummy 3-vertex grid.
-
-        Corner nodes case.
-        """
-        # single triangle with point on antimeridian
-        source_verts = np.array([(0.0, 90.0), (-180, 0.0), (0.0, -90)])
-        source_data_single_dim = [1.0, 2.0, 3.0]
-        source_grid = ux.open_grid(source_verts)
-        destination_grid = ux.open_grid(source_verts)
-
-        destination_single_data = _nearest_neighbor(source_grid,
-                                                    destination_grid,
-                                                    source_data_single_dim,
-                                                    remap_to="nodes")
-
-        source_data_multi_dim = np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0],
-                                          [7.0, 8.0, 9.0]])
-
-        destination_multi_data = _nearest_neighbor(source_grid,
-                                                   destination_grid,
-                                                   source_data_multi_dim,
-                                                   remap_to="nodes")
-
-        assert np.array_equal(source_data_single_dim, destination_single_data)
-        assert np.array_equal(source_data_multi_dim, destination_multi_data)
-
-    def test_remap_to_corner_nodes_cartesian(self):
-        """Test remapping to the same dummy 3-vertex grid, using cartesian
-        coordinates.
-
-        Corner nodes case.
-        """
+def test_remap_to_same_grid_corner_nodes():
+    """Test remapping to the same dummy 3-vertex grid. Corner nodes case."""
+    source_verts = np.array([(0.0, 90.0), (-180, 0.0), (0.0, -90)])
+    source_data_single_dim = [1.0, 2.0, 3.0]
+    source_grid = ux.open_grid(source_verts)
+    destination_grid = ux.open_grid(source_verts)
 
-        # single triangle
-        source_verts = np.array([(0.0, 0.0, 1.0), (0.0, 1.0, 0.0),
-                                 (1.0, 0.0, 0.0)])
-        source_data_single_dim = [1.0, 2.0, 3.0]
+    destination_single_data = _nearest_neighbor(source_grid,
+                                                destination_grid,
+                                                source_data_single_dim,
+                                                remap_to="nodes")
 
-        # open the source and destination grids
-        source_grid = ux.open_grid(source_verts)
-        destination_grid = ux.open_grid(source_verts)
+    source_data_multi_dim = np.array([[1.0, 2.0, 3.0], [4.0, 5.0, 6.0],
+                                       [7.0, 8.0, 9.0]])
 
-        # create the destination data using the nearest neighbor function
-        destination_data = _nearest_neighbor(source_grid,
-                                             destination_grid,
-                                             source_data_single_dim,
-                                             remap_to="nodes",
-                                             coord_type="cartesian")
+    destination_multi_data = _nearest_neighbor(source_grid,
+                                               destination_grid,
+                                               source_data_multi_dim,
+                                               remap_to="nodes")
 
-        # assert that the source and destination data are the same
-        assert np.array_equal(source_data_single_dim, destination_data)
+    nt.assert_array_equal(source_data_single_dim, destination_single_data)
+    nt.assert_array_equal(source_data_multi_dim, destination_multi_data)
 
-    def test_nn_remap(self):
-        """Test nearest neighbor remapping.
 
-        Steps:
-        1. Open a grid and a dataset,
-        2. Open the grid to remap dataset in 1
-        3. Remap the dataset in 1 to the grid in 2
-        """
-        uxds = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+def test_remap_to_corner_nodes_cartesian():
+    """Test remapping to the same dummy 3-vertex grid, using cartesian coordinates. Corner nodes case."""
+    source_verts = np.array([(0.0, 0.0, 1.0), (0.0, 1.0, 0.0), (1.0, 0.0, 0.0)])
+    source_data_single_dim = [1.0, 2.0, 3.0]
+    source_grid = ux.open_grid(source_verts)
+    destination_grid = ux.open_grid(source_verts)
 
-        uxgrid = ux.open_grid(gridfile_ne30)
+    destination_data = _nearest_neighbor(source_grid,
+                                         destination_grid,
+                                         source_data_single_dim,
+                                         remap_to="nodes",
+                                         coord_type="cartesian")
 
-        uxda = uxds['v1']
-        out_da = uxda.remap.nearest_neighbor(destination_obj=uxgrid, remap_to="nodes")
+    nt.assert_array_equal(source_data_single_dim, destination_data)
 
-        # Assert the remapping was successful and the variable is populated
-        self.assertTrue(len(out_da) != 0)
 
-    def test_remap_return_types(self):
-        """Tests the return type of the `UxDataset` and `UxDataArray`
-        implementations of Nearest Neighbor Remapping."""
-        source_data_paths = [
-            dsfile_v1_geoflow, dsfile_v2_geoflow, dsfile_v3_geoflow
-        ]
-        source_uxds = ux.open_mfdataset(gridfile_geoflow, source_data_paths)
-        destination_uxds = ux.open_dataset(gridfile_CSne30,
-                                           dsfile_vortex_CSne30)
+def test_nn_remap():
+    """Test nearest neighbor remapping."""
+    uxds = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    uxgrid = ux.open_grid(gridfile_ne30)
+    uxda = uxds['v1']
+    out_da = uxda.remap.nearest_neighbor(destination_grid=uxgrid, remap_to="nodes")
 
-        remap_uxda_to_grid = source_uxds['v1'].remap.nearest_neighbor(
-            destination_uxds.uxgrid)
+    assert len(out_da) != 0
 
-        assert isinstance(remap_uxda_to_grid, UxDataArray)
 
-        remap_uxda_to_uxda = source_uxds['v1'].remap.nearest_neighbor(
-            destination_uxds['psi'])
+def test_remap_return_types():
+    """Tests the return type of the `UxDataset` and `UxDataArray` implementations of Nearest Neighbor Remapping."""
+    source_data_paths = [dsfile_v1_geoflow, dsfile_v2_geoflow, dsfile_v3_geoflow]
+    source_uxds = ux.open_mfdataset(gridfile_geoflow, source_data_paths)
+    destination_grid = ux.open_grid(gridfile_CSne30)
 
-        # Dataset with two vars: original "psi" and remapped "v1"
-        assert isinstance(remap_uxda_to_uxda, UxDataset)
-        assert len(remap_uxda_to_uxda.data_vars) == 2
+    remap_uxda_to_grid = source_uxds['v1'].remap.nearest_neighbor(destination_grid)
 
-        remap_uxda_to_uxds = source_uxds['v1'].remap.nearest_neighbor(
-            destination_uxds)
+    assert isinstance(remap_uxda_to_grid, UxDataArray)
 
-        # Dataset with two vars: original "psi" and remapped "v1"
-        assert isinstance(remap_uxda_to_uxds, UxDataset)
-        assert len(remap_uxda_to_uxds.data_vars) == 2
+    remap_uxds_to_grid = source_uxds.remap.nearest_neighbor(destination_grid)
 
-        remap_uxds_to_grid = source_uxds.remap.nearest_neighbor(
-            destination_uxds.uxgrid)
+    assert isinstance(remap_uxds_to_grid, UxDataset)
+    assert len(remap_uxds_to_grid.data_vars) == 3
 
-        # Dataset with three vars: remapped "v1, v2, v3"
-        assert isinstance(remap_uxds_to_grid, UxDataset)
-        assert len(remap_uxds_to_grid.data_vars) == 3
 
-        remap_uxds_to_uxda = source_uxds.remap.nearest_neighbor(
-            destination_uxds['psi'])
+def test_edge_centers_remapping():
+    """Tests the ability to remap on edge centers using Nearest Neighbor Remapping."""
+    source_grid = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    destination_grid = ux.open_grid(mpasfile_QU)
 
-        # Dataset with four vars: original "psi" and remapped "v1, v2, v3"
-        assert isinstance(remap_uxds_to_uxda, UxDataset)
-        assert len(remap_uxds_to_uxda.data_vars) == 4
+    remap_to_edge_centers_spherical = source_grid['v1'].remap.nearest_neighbor(destination_grid=destination_grid,
+                                                                         remap_to="edge centers", coord_type='spherical')
 
-        remap_uxds_to_uxds = source_uxds.remap.nearest_neighbor(
-            destination_uxds)
+    remap_to_edge_centers_cartesian = source_grid['v1'].remap.nearest_neighbor(destination_grid=destination_grid,
+                                                                         remap_to="edge centers", coord_type='cartesian')
 
-        # Dataset with four vars: original "psi" and remapped "v1, v2, v3"
-        assert isinstance(remap_uxds_to_uxds, UxDataset)
-        assert len(remap_uxds_to_uxds.data_vars) == 4
+    assert remap_to_edge_centers_spherical._edge_centered()
+    assert remap_to_edge_centers_cartesian._edge_centered()
 
-    def test_edge_centers_remapping(self):
-        """Tests the ability to remap on edge centers using Nearest Neighbor
-        Remapping."""
 
-        # Open source and destination datasets to remap to
-        source_grid = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
-        destination_grid = ux.open_dataset(mpasfile_QU, mpasfile_QU)
+def test_overwrite():
+    """Tests that the remapping no longer overwrites the dataset."""
+    source_grid = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    destination_dataset = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
 
-        remap_to_edge_centers = source_grid['v1'].remap.nearest_neighbor(destination_obj=destination_grid,
-                                                                         remap_to="edge centers")
+    remap_to_edge_centers = source_grid['v1'].remap.nearest_neighbor(destination_grid=destination_dataset.uxgrid,
+                                                                     remap_to="face centers", coord_type='cartesian')
 
-        # Assert the data variable lies on the "edge centers"
-        self.assertTrue(destination_grid['v1']._edge_centered())
+    assert not np.array_equal(destination_dataset['v1'], remap_to_edge_centers)
 
 
-class TestInverseDistanceWeightedRemapping(TestCase):
-    """Testing for inverse distance weighted remapping."""
+def test_source_data_remap():
+    """Test the remapping of all source data positions."""
+    source_uxds = ux.open_dataset(mpasfile_QU, mpasfile_QU)
+    destination_grid = ux.open_grid(gridfile_geoflow)
 
-    def test_remap_center_nodes(self):
-        """Test remapping to center nodes."""
+    face_centers = source_uxds['latCell'].remap.nearest_neighbor(destination_grid=destination_grid, remap_to="nodes")
+    nodes = source_uxds['latVertex'].remap.nearest_neighbor(destination_grid=destination_grid, remap_to="nodes")
+    edges = source_uxds['angleEdge'].remap.nearest_neighbor(destination_grid=destination_grid, remap_to="nodes")
 
-        # datasets to use for remap
-        dataset = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
-        destination_uxds = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    assert len(face_centers.values) != 0
+    assert len(nodes.values) != 0
+    assert len(edges.values) != 0
 
-        data_on_face_centers = dataset['v1'].remap.inverse_distance_weighted(
-            destination_uxds.uxgrid, remap_to="face centers")
 
-        assert not np.array_equal(dataset['v1'], data_on_face_centers)
+def test_value_errors():
+    """Tests the raising of value errors and warnings in the function."""
+    source_uxds = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    source_uxds_2 = ux.open_dataset(mpasfile_QU, mpasfile_QU)
+    destination_grid = ux.open_grid(gridfile_geoflow)
 
-    def test_remap_nodes(self):
-        """Test remapping to nodes."""
+    with nt.assert_raises(ValueError):
+        source_uxds['v1'].remap.nearest_neighbor(destination_grid=destination_grid, remap_to="test", coord_type='spherical')
+    with nt.assert_raises(ValueError):
+        source_uxds['v1'].remap.nearest_neighbor(destination_grid=destination_grid, remap_to="test", coord_type="cartesian")
+    with nt.assert_raises(ValueError):
+        source_uxds['v1'].remap.nearest_neighbor(destination_grid=destination_grid, remap_to="nodes", coord_type="test")
+    with nt.assert_raises(ValueError):
+        source_uxds_2['cellsOnCell'].remap.nearest_neighbor(destination_grid=destination_grid, remap_to="nodes")
 
-        # datasets to use for remap
-        dataset = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
-        destination_uxds = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+def test_preserve_coordinates():
+    """Tests if coordinates are preserved after remapping."""
+    source_uxds = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    destination_grid = ux.open_grid(mpasfile_QU)
 
-        data_on_nodes = dataset['v1'].remap.inverse_distance_weighted(
-            destination_uxds.uxgrid, remap_to="nodes")
+    res = source_uxds.remap.nearest_neighbor(destination_grid=destination_grid)
 
-        assert not np.array_equal(dataset['v1'], data_on_nodes)
+    assert "time" in res.coords
 
-    def test_cartesian_remap_to_nodes(self):
-        """Test remapping using cartesian coordinates using nodes."""
 
-        # triangle with data on nodes
-        source_verts = np.array([(0.0, 0.0, 1.0), (0.0, 1.0, 0.0),
-                                 (1.0, 0.0, 0.0)])
-        source_data = [1.0, 2.0, 3.0]
+# Inverse Distance Weighted Remapping Tests
+def test_remap_center_nodes():
+    """Test remapping to center nodes."""
+    dataset = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    destination_grid = ux.open_grid(gridfile_geoflow)
 
-        # open the source and destination grids
-        source_grid = ux.open_grid(source_verts)
-        destination_grid = ux.open_grid(source_verts)
+    data_on_face_centers = dataset['v1'].remap.inverse_distance_weighted(destination_grid, remap_to="face centers", power=6)
 
-        # create the first destination data using two k neighbors
-        destination_data_neighbors_2 = _inverse_distance_weighted_remap(
-            source_grid,
-            destination_grid,
-            source_data,
-            remap_to="nodes",
-            coord_type="cartesian",
-            k=3)
+    assert not np.array_equal(dataset['v1'], data_on_face_centers)
 
-        # create the second destination data using one k neighbor
-        destination_data_neighbors_1 = _inverse_distance_weighted_remap(
-            source_grid,
-            destination_grid,
-            source_data,
-            remap_to="nodes",
-            coord_type="cartesian",
-            k=2)
 
-        # two different k remaps are different
-        assert not np.array_equal(destination_data_neighbors_1,
-                                  destination_data_neighbors_2)
+def test_remap_nodes():
+    """Test remapping to nodes."""
+    dataset = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    destination_grid = ux.open_grid(gridfile_geoflow)
 
-    def test_remap_return_types(self):
-        """Tests the return type of the `UxDataset` and `UxDataArray`
-        implementations of Inverse Distance Weighted."""
+    data_on_nodes = dataset['v1'].remap.inverse_distance_weighted(destination_grid, remap_to="nodes")
 
-        source_data_paths = [
-            dsfile_v1_geoflow, dsfile_v2_geoflow, dsfile_v3_geoflow
-        ]
-        source_uxds = ux.open_mfdataset(gridfile_geoflow, source_data_paths)
-        destination_uxds = ux.open_dataset(gridfile_CSne30,
-                                           dsfile_vortex_CSne30)
+    assert not np.array_equal(dataset['v1'], data_on_nodes)
 
-        remap_uxda_to_grid = source_uxds['v1'].remap.inverse_distance_weighted(
-            destination_uxds.uxgrid, power=3, k=10)
 
-        assert isinstance(remap_uxda_to_grid, UxDataArray)
-        assert len(remap_uxda_to_grid) == 1
+def test_cartesian_remap_to_nodes():
+    """Test remapping using cartesian coordinates."""
+    source_verts = np.array([(0.0, 0.0, 1.0), (0.0, 1.0, 0.0), (1.0, 0.0, 0.0)])
+    source_data = [1.0, 2.0, 3.0]
+    source_grid = ux.open_grid(source_verts)
+    destination_grid = ux.open_grid(source_verts)
 
-        remap_uxda_to_uxda = source_uxds['v1'].remap.inverse_distance_weighted(
-            destination_uxds['psi'], power=3, k=10)
+    destination_data_neighbors_2 = _inverse_distance_weighted_remap(source_grid, destination_grid, source_data, remap_to="nodes", coord_type="cartesian", k=3)
+    destination_data_neighbors_1 = _inverse_distance_weighted_remap(source_grid, destination_grid, source_data, remap_to="nodes", coord_type="cartesian", k=2)
 
-        # Dataset with two vars: original "psi" and remapped "v1"
-        assert isinstance(remap_uxda_to_uxda, UxDataset)
-        assert len(remap_uxda_to_uxda.data_vars) == 2
+    assert not np.array_equal(destination_data_neighbors_1, destination_data_neighbors_2)
 
-        remap_uxda_to_uxds = source_uxds['v1'].remap.inverse_distance_weighted(
-            destination_uxds, power=3, k=10)
 
-        # Dataset with two vars: original "psi" and remapped "v1"
-        assert isinstance(remap_uxda_to_uxds, UxDataset)
-        assert len(remap_uxda_to_uxds.data_vars) == 2
+def test_remap_return_types_idw():
+    """Tests the return type of the `UxDataset` and `UxDataArray` implementations of Inverse Distance Weighted."""
+    source_data_paths = [dsfile_v1_geoflow, dsfile_v2_geoflow, dsfile_v3_geoflow]
+    source_uxds = ux.open_mfdataset(gridfile_geoflow, source_data_paths)
+    destination_grid = ux.open_grid(gridfile_CSne30)
 
-        remap_uxds_to_grid = source_uxds.remap.inverse_distance_weighted(
-            destination_uxds.uxgrid)
+    remap_uxda_to_grid = source_uxds['v1'].remap.inverse_distance_weighted(destination_grid, power=3, k=10)
 
-        # Dataset with three vars: remapped "v1, v2, v3"
-        assert isinstance(remap_uxds_to_grid, UxDataset)
-        assert len(remap_uxds_to_grid.data_vars) == 3
+    assert isinstance(remap_uxda_to_grid, UxDataArray)
+    assert len(remap_uxda_to_grid) == 1
 
-        remap_uxds_to_uxda = source_uxds.remap.inverse_distance_weighted(
-            destination_uxds['psi'])
+    remap_uxds_to_grid = source_uxds.remap.inverse_distance_weighted(destination_grid)
 
-        # Dataset with four vars: original "psi" and remapped "v1, v2, v3"
-        assert isinstance(remap_uxds_to_uxda, UxDataset)
-        assert len(remap_uxds_to_uxda.data_vars) == 4
+    assert isinstance(remap_uxds_to_grid, UxDataset)
+    assert len(remap_uxds_to_grid.data_vars) == 3
 
-        remap_uxds_to_uxds = source_uxds.remap.inverse_distance_weighted(
-            destination_uxds)
 
-        # Dataset with four vars: original "psi" and remapped "v1, v2, v3"
-        assert isinstance(remap_uxds_to_uxds, UxDataset)
-        assert len(remap_uxds_to_uxds.data_vars) == 4
+def test_edge_remapping():
+    """Tests the ability to remap on edge centers using Inverse Distance Weighted Remapping."""
+    source_grid = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    destination_grid = ux.open_grid(mpasfile_QU)
 
-    def test_edge_remapping(self):
-        """Tests the ability to remap on edge centers using Inverse Distance
-        Weighted Remapping."""
+    remap_to_edge_centers_spherical = source_grid['v1'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="edge centers", coord_type='spherical')
+    remap_to_edge_centers_cartesian = source_grid['v1'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="edge centers", coord_type='cartesian')
 
-        # Open source and destination datasets to remap to
-        source_grid = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
-        destination_grid = ux.open_dataset(mpasfile_QU, mpasfile_QU)
+    assert remap_to_edge_centers_spherical._edge_centered()
+    assert remap_to_edge_centers_cartesian._edge_centered()
 
-        # Perform remapping to the edge centers of the dataset
 
-        remap_to_edge_centers = source_grid['v1'].remap.inverse_distance_weighted(destination_obj=destination_grid,
-                                                                                  remap_to="edge centers")
+def test_overwrite_idw():
+    """Tests that the remapping no longer overwrites the dataset."""
+    source_grid = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    destination_dataset = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
 
-        # Assert the data variable lies on the "edge centers"
-        self.assertTrue(destination_grid['v1']._edge_centered())
+    remap_to_edge_centers = source_grid['v1'].remap.inverse_distance_weighted(destination_grid=destination_dataset.uxgrid, remap_to="face centers", coord_type='cartesian')
+
+    assert not np.array_equal(destination_dataset['v1'], remap_to_edge_centers)
+
+
+def test_source_data_remap_idw():
+    """Test the remapping of all source data positions."""
+    source_uxds = ux.open_dataset(mpasfile_QU, mpasfile_QU)
+    destination_grid = ux.open_grid(gridfile_geoflow)
+
+    face_centers = source_uxds['latCell'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="nodes")
+    nodes = source_uxds['latVertex'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="nodes")
+    edges = source_uxds['angleEdge'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="nodes")
+
+    assert len(face_centers.values) != 0
+    assert len(nodes.values) != 0
+    assert len(edges.values) != 0
+
+
+def test_value_errors_idw():
+    """Tests the raising of value errors and warnings in the function."""
+    source_uxds = ux.open_dataset(gridfile_geoflow, dsfile_v1_geoflow)
+    source_uxds_2 = ux.open_dataset(mpasfile_QU, mpasfile_QU)
+    destination_grid = ux.open_grid(gridfile_geoflow)
+
+    with nt.assert_raises(ValueError):
+        source_uxds['v1'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="nodes", k=1)
+
+    with nt.assert_raises(ValueError):
+        source_uxds['v1'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="nodes", k=source_uxds.uxgrid.n_node + 1)
+
+    with nt.assert_raises(ValueError):
+        source_uxds['v1'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="test", k=2, coord_type='spherical')
+    with nt.assert_raises(ValueError):
+        source_uxds['v1'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="test", k=2, coord_type="cartesian")
+
+    with nt.assert_raises(ValueError):
+        source_uxds['v1'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="nodes", k=2, coord_type="test")
+
+    with nt.assert_raises(ValueError):
+        source_uxds_2['cellsOnCell'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="nodes")
+
+    with nt.assert_warns(UserWarning):
+        source_uxds['v1'].remap.inverse_distance_weighted(destination_grid=destination_grid, remap_to="nodes", power=6)
