@@ -10,9 +10,14 @@ INT_FILL_VALUE = np.iinfo(INT_DTYPE).min
 # half of the working precision is already the most optimal value for the error tolerance,
 # more tailored values will be used in the future.
 
-ERROR_TOLERANCE = 1.0e-8
+ERROR_TOLERANCE = np.float64(1.0e-8)
 
-ENABLE_JIT_CACHE = True
-ENABLE_JIT = True
+# The below value is the machine epsilon for the float64 data type, it will be used in the most basic operations as a
+# error tolerance, mainly in the intersection calculations.
+MACHINE_EPSILON = np.float64(np.finfo(float).eps)
+
+ENABLE_FMA = False
 
 GRID_DIMS = ["n_node", "n_edge", "n_face"]
+
+WGS84_CRS = "EPSG:4326"
