@@ -1,28 +1,31 @@
 from __future__ import annotations
 
+import os
+import sys
+from html import escape
+from typing import IO, Optional, Union, Any
+from warnings import warn
+
 import numpy as np
 import xarray as xr
-
-import sys
-
-from typing import Optional, IO, Union
-
-from uxarray.grid import Grid
-from uxarray.core.dataarray import UxDataArray
-
-from uxarray.plot.accessor import UxDatasetPlotAccessor
-
+from xarray.core import dtypes
+from xarray.core.options import OPTIONS
 from xarray.core.utils import UncachedAccessor
 
+import uxarray
+from uxarray.core.dataarray import UxDataArray
+from uxarray.core.utils import _map_dims_to_ugrid
+from uxarray.formatting_html import dataset_repr
+from uxarray.grid import Grid
+from uxarray.grid.dual import construct_dual
+from uxarray.grid.validation import _check_duplicate_nodes_indices
+from uxarray.plot.accessor import UxDatasetPlotAccessor
 from uxarray.remap import UxDatasetRemapAccessor
-
-from warnings import warn
 
 
 class UxDataset(xr.Dataset):
-    """A ``xarray.Dataset``-like, multi-dimensional, in memory, array database.
-    Inherits from ``xarray.Dataset`` and has its own unstructured grid-aware
-    dataset operators and attributes through the ``uxgrid`` accessor.
+    """Grid informed ``xarray.Dataset`` with an attached ``Grid`` accessor and
+    grid-specific functionality.
 
     Parameters
     ----------
@@ -68,13 +71,18 @@ class UxDataset(xr.Dataset):
                 "an instance of the `uxarray.Grid` class"
             )
         else:
-            self.uxgrid = uxgrid
+            self._uxgrid = uxgrid
 
         super().__init__(*args, **kwargs)
 
     # declare plotting accessor
     plot = UncachedAccessor(UxDatasetPlotAccessor)
     remap = UncachedAccessor(UxDatasetRemapAccessor)
+
+    def _repr_html_(self) -> str:
+        if OPTIONS["display_style"] == "text":
+            return f"<pre>{escape(repr(self))}</pre>"
+        return dataset_repr(self)
 
     def __getitem__(self, key):
         """Override to make sure the result is an instance of
@@ -123,14 +131,8 @@ class UxDataset(xr.Dataset):
 
     @property
     def uxgrid(self):
-        """``uxarray.Grid`` property for ``uxarray.UxDataset`` to make it
-        unstructured grid-aware.
-
-        Examples
-        --------
-        uxds = ux.open_dataset(grid_path, data_path)
-        uxds.uxgrid
-        """
+        """Linked ``Grid`` representing to the unstructured grid the data
+        resides on."""
         return self._uxgrid
 
     # a setter function
@@ -210,6 +212,105 @@ class UxDataset(xr.Dataset):
             coords={"index": range(len(next(iter(data.values()))))},
             **kwargs,
         )
+
+    @classmethod
+    def from_structured(cls, ds: xr.Dataset):
+        """Converts a structured ``xarray.Dataset`` into an unstructured ``uxarray.UxDataset``
+
+        Parameters
+        ----------
+        ds : xr.Dataset
+            The structured `xarray.Dataset` to convert. Must contain longitude and latitude variables consistent
+            with the CF-conventions
+
+        tol : float, optional
+            Tolerance for considering nodes as identical when constructing the grid from longitude and latitude.
+            Default is `1e-10`.
+
+        Returns
+        -------
+        UxDataset
+            An instance of `uxarray.UxDataset`
+        """
+        from uxarray import Grid
+
+        uxgrid = Grid.from_dataset(ds)
+
+        ds = _map_dims_to_ugrid(ds, uxgrid._source_dims_dict, uxgrid)
+
+        # Drop spatial coordinates
+        coords_to_drop = [
+            coord for coord, da_coord in ds.coords.items() if "n_face" in da_coord.dims
+        ]
+        ds = ds.drop_vars(coords_to_drop)
+
+        return cls(ds, uxgrid=uxgrid)
+
+    @classmethod
+    def from_xarray(cls, ds: xr.Dataset, uxgrid: Grid = None, ugrid_dims: dict = None):
+        """
+        Converts a ``xarray.Dataset`` into a ``uxarray.UxDataset``, paired with either a user-defined or
+        parsed ``Grid``
+
+        Parameters
+        ----------
+        ds: xr.Dataset
+            An Xarray dataset containing data residing on an unstructured grid
+        uxgrid: Grid, optional
+            ``Grid`` object representing an unstructured grid. If a grid is not provided, the source ds will be
+            parsed to see if a ``Grid`` can be constructed.
+        ugrid_dims: dict, optional
+            A dictionary mapping dataset dimensions to UGRID dimensions.
+
+        Returns
+        -------
+        cls
+            A ``ux.UxDataset`` with data from the ``xr.Dataset` paired with a ``ux.Grid``
+        """
+        if uxgrid is not None:
+            if ugrid_dims is None and uxgrid._source_dims_dict is not None:
+                ugrid_dims = uxgrid._source_dims_dict
+            # Grid is provided,
+        else:
+            # parse
+            uxgrid = Grid.from_dataset(ds)
+            ugrid_dims = uxgrid._source_dims_dict
+
+        # map each dimension to its UGRID equivalent
+        ds = _map_dims_to_ugrid(ds, ugrid_dims, uxgrid)
+
+        return cls(ds, uxgrid=uxgrid)
+
+    @classmethod
+    def from_healpix(cls, ds: Union[str, os.PathLike, xr.Dataset], **kwargs):
+        """
+        Loads a dataset represented in the HEALPix format into a ``ux.UxDataSet``, paired
+        with a ``Grid`` containing information about the HEALPix definition.
+
+        Parameters
+        ----------
+        ds: str, os.PathLike, xr.Dataset
+            Reference to a HEALPix Dataset
+
+        Returns
+        -------
+        cls
+            A ``ux.UxDataset`` instance
+        """
+
+        if not isinstance(ds, xr.Dataset):
+            ds = xr.open_dataset(ds, **kwargs)
+
+        if "cell" not in ds.dims:
+            raise ValueError("Healpix dataset must contain a 'cell' dimension.")
+
+        # Compute the HEALPix Zoom Level
+        zoom = np.emath.logn(4, (ds.sizes["cell"] / 12)).astype(int)
+
+        # Attach a  HEALPix Grid
+        uxgrid = Grid.from_healpix(zoom)
+
+        return cls.from_xarray(ds, uxgrid, {"cell": "n_face"})
 
     def info(self, buf: IO = None, show_attrs=False) -> None:
         """Concise summary of Dataset variables and attributes including grid
@@ -327,61 +428,93 @@ class UxDataset(xr.Dataset):
         xarr = super().to_array()
         return UxDataArray(xarr, uxgrid=self.uxgrid)
 
-    def nearest_neighbor_remap(
-        self,
-        destination_obj: Union[Grid, UxDataArray, UxDataset],
-        remap_to: str = "nodes",
-        coord_type: str = "spherical",
-    ):
-        """Nearest Neighbor Remapping between a source (``UxDataset``) and
-        destination.`.
+    def to_xarray(self, grid_format: str = "UGRID") -> xr.Dataset:
+        """
+        Converts a ``ux.UXDataset`` to a ``xr.Dataset``.
 
         Parameters
-        ---------
-        destination_obj : Grid, UxDataArray, UxDataset
-            Destination for remapping
-        remap_to : str, default="nodes"
-            Location of where to map data, either "nodes", "edge centers", or "face centers"
-        coord_type : str, default="spherical"
-            Indicates whether to remap using on spherical or cartesian coordinates
+        ----------
+        grid_format : str, default="UGRID"
+            The format in which to convert the grid. Supported values are "UGRID" and "HEALPix". The dimensions will
+            match the selected grid format.
+
+        Returns
+        -------
+        xr.Dataset
+            The ``ux.UXDataset`` represented as a ``xr.Dataset``
         """
-        warn(
-            "This usage of remapping will be deprecated in a future release. It is advised to use uxds.remap.nearest_neighbor() instead.",
-            DeprecationWarning,
+        if grid_format == "HEALPix":
+            ds = self.rename_dims({"n_face": "cell"})
+            return xr.Dataset(ds)
+
+        return xr.Dataset(self)
+
+    def get_dual(self):
+        """Compute the dual mesh for a dataset, returns a new dataset object.
+
+        Returns
+        --------
+        dual : uxds
+            Dual Mesh `uxds` constructed
+        """
+
+        if _check_duplicate_nodes_indices(self.uxgrid):
+            raise RuntimeError("Duplicate nodes found, cannot construct dual")
+
+        if self.uxgrid.partial_sphere_coverage:
+            warn(
+                "This mesh is partial, which could cause inconsistent results and data will be lost",
+                Warning,
+            )
+
+        # Get dual mesh node face connectivity
+        dual_node_face_conn = construct_dual(grid=self.uxgrid)
+
+        # Construct dual mesh
+        dual = self.uxgrid.from_topology(
+            self.uxgrid.face_lon.values,
+            self.uxgrid.face_lat.values,
+            dual_node_face_conn,
         )
 
-        return self.remap.nearest_neighbor(destination_obj, remap_to, coord_type)
+        # Initialize new dataset
+        dataset = uxarray.UxDataset(uxgrid=dual)
 
-    def inverse_distance_weighted_remap(
-        self,
-        destination_obj: Union[Grid, UxDataArray, UxDataset],
-        remap_to: str = "nodes",
-        coord_type: str = "spherical",
-        power=2,
-        k=8,
+        # Dictionary to swap dimensions
+        dim_map = {"n_face": "n_node", "n_node": "n_face"}
+
+        # For each data array in the dataset, reconstruct the data array with the dual mesh
+        for var in self.data_vars:
+            # Get correct dimensions for the dual
+            dims = [dim_map.get(dim, dim) for dim in self[var].dims]
+
+            # Get the values from the data array
+            data = np.array(self[var].values)
+
+            # Construct the new data array
+            uxda = uxarray.UxDataArray(uxgrid=dual, data=data, dims=dims, name=var)
+
+            # Add data array to dataset
+            dataset[var] = uxda
+
+        return dataset
+
+    def where(self, cond: Any, other: Any = dtypes.NA, drop: bool = False):
+        return UxDataset(self.to_xarray().where(cond, other, drop), uxgrid=self.uxgrid)
+
+    where.__doc__ = xr.Dataset.where.__doc__
+
+    def sel(
+        self, indexers=None, method=None, tolerance=None, drop=False, **indexers_kwargs
     ):
-        """Inverse Distance Weighted Remapping between a source (``UxDataset``)
-        and destination.`.
-
-        Parameters
-        ---------
-        destination_obj : Grid, UxDataArray, UxDataset
-            Destination for remapping
-        remap_to : str, default="nodes"
-            Location of where to map data, either "nodes", "edge centers", or "face centers"
-        coord_type : str, default="spherical"
-            Indicates whether to remap using on spherical or cartesian coordinates
-        power : int, default=2
-            Power parameter for inverse distance weighting. This controls how local or global the remapping is, a higher
-            power causes points that are further away to have less influence
-        k : int, default=8
-            Number of nearest neighbors to consider in the weighted calculation.
-        """
-        warn(
-            "This usage of remapping will be deprecated in a future release. It is advised to use uxds.remap.inverse_distance_weighted() instead.",
-            DeprecationWarning,
+        return UxDataset(
+            self.to_xarray().sel(indexers, tolerance, drop, **indexers_kwargs),
+            uxgrid=self.uxgrid,
         )
 
-        return self.remap.inverse_distance_weighted(
-            destination_obj, remap_to, coord_type, power, k
-        )
+    sel.__doc__ = xr.Dataset.sel.__doc__
+
+    def fillna(self, value: Any):
+        return UxDataset(super().fillna(value), uxgrid=self.uxgrid)
+
+    fillna.__doc__ = xr.Dataset.fillna.__doc__
