@@ -1,9 +1,18 @@
 import numpy as np
+import math
 
-# from uxarray.grid.coordinates import node_xyz_to_lonlat_rad, normalize_in_place
 
-from uxarray.grid.coordinates import _xyz_to_lonlat_rad, _normalize_xyz
-from uxarray.constants import ERROR_TOLERANCE
+from uxarray.grid.coordinates import (
+    _normalize_xyz_scalar,
+)
+
+from uxarray.grid.utils import _angle_of_2_vectors
+
+from uxarray.constants import ERROR_TOLERANCE, MACHINE_EPSILON
+
+from uxarray.utils.computing import isclose, dot
+
+from numba import njit
 
 
 def _to_list(obj):
@@ -17,158 +26,71 @@ def _to_list(obj):
     return obj
 
 
-def point_within_gca(pt, gca_cart, is_directed=False):
-    """Check if a point lies on a given Great Circle Arc (GCA). The anti-
-    meridian case is also considered.
+@njit(cache=True)
+def point_within_gca(pt_xyz, gca_a_xyz, gca_b_xyz):
+    """
+    Check if a point lies on a given Great Circle Arc (GCA) interval, considering the smaller arc of the circle.
+    Handles the anti-meridian case as well.
 
     Parameters
     ----------
-    pt : numpy.ndarray (float)
+    pt_xyz : numpy.ndarray
         Cartesian coordinates of the point.
-    gca_cart : numpy.ndarray of shape (2, 3), (np.float or gmpy2.mpfr)
-        Cartesian coordinates of the Great Circle Arc (GCR).
-    is_directed : bool, optional, default = False
-        If True, the GCA is considered to be directed, which means it can only from v0-->v1. If False, the GCA is undirected,
-        and we will always assume the small circle (The one less than 180 degree) side is the GCA. The default is False.
-        For the case of the anti-podal case, the direction is v_0--> the pole point that on the same hemisphere as v_0-->v_1
+    gca_a_xyz : numpy.ndarray
+        Cartesian coordinates of the first endpoint of the Great Circle Arc.
+    gca_b_xyz : numpy.ndarray
+        Cartesian coordinates of the second endpoint of the Great Circle Arc.
 
     Returns
     -------
     bool
-        True if the point lies between the two endpoints of the GCR, False otherwise.
+        True if the point lies within the specified GCA interval, False otherwise.
 
     Raises
     ------
     ValueError
-        If the input GCR spans exactly 180 degrees (π radians), as this GCR can have multiple planes.
-        In such cases, consider breaking the GCR into two separate GCRs.
-
-    ValueError
-        If the input GCR spans more than 180 degrees (π radians).
-        In such cases, consider breaking the GCR into two separate GCRs.
+        If the input GCA spans exactly 180 degrees (π radians), as this GCA can have multiple planes.
+        In such cases, consider breaking the GCA into two separate arcs.
 
     Notes
     -----
-    The function checks if the given point is on the Great Circle Arc by considering its cartesian coordinates and
-    accounting for the anti-meridian case.
-
-    The anti-meridian case occurs when the GCR crosses the anti-meridian (0 longitude).
-    In this case, the function handles scenarios where the GCA spans across more than 180 degrees, requiring specific operation.
-
-    The function relies on the `_angle_of_2_vectors` and `is_between` functions to perform the necessary calculations.
-
-    Please ensure that the input coordinates are in radians and adhere to the ERROR_TOLERANCE value for floating-point comparisons.
+    - The function ensures that the point lies on the same plane as the GCA before performing interval checks.
+    - It assumes the input represents the smaller arc of the Great Circle.
+    - The `_angle_of_2_vectors` and `_xyz_to_lonlat_rad_scalar` functions are used for calculations.
     """
-    # Convert the cartesian coordinates to lonlat coordinates
-    pt_lonlat = np.array(_xyz_to_lonlat_rad(pt[0], pt[1], pt[2]))
-    GCRv0_lonlat = np.array(
-        _xyz_to_lonlat_rad(gca_cart[0][0], gca_cart[0][1], gca_cart[0][2])
-    )
-    GCRv1_lonlat = np.array(
-        _xyz_to_lonlat_rad(gca_cart[1][0], gca_cart[1][1], gca_cart[1][2])
-    )
-
-    # Convert the list to np.float64
-    gca_cart[0] = np.array(gca_cart[0], dtype=np.float64)
-    gca_cart[1] = np.array(gca_cart[1], dtype=np.float64)
-
-    # First if the input GCR is exactly 180 degree, we throw an exception, since this GCR can have multiple planes
-    angle = _angle_of_2_vectors(gca_cart[0], gca_cart[1])
-    if np.allclose(angle, np.pi, rtol=0, atol=ERROR_TOLERANCE):
+    # 1. Check if the input GCA spans exactly 180 degrees
+    angle_ab = _angle_of_2_vectors(gca_a_xyz, gca_b_xyz)
+    if np.allclose(angle_ab, np.pi, rtol=0.0, atol=MACHINE_EPSILON):
         raise ValueError(
-            "The input Great Circle Arc is exactly 180 degree, this Great Circle Arc can have multiple planes. "
-            "Consider breaking the Great Circle Arc"
-            "into two Great Circle Arcs"
+            "The input Great Circle Arc spans exactly 180 degrees, which can correspond to multiple planes. "
+            "Consider breaking the Great Circle Arc into two smaller arcs."
         )
 
+    # 2. Verify if the point lies on the plane of the GCA
+    cross_product = np.cross(gca_a_xyz, gca_b_xyz)
     if not np.allclose(
-        np.dot(np.cross(gca_cart[0], gca_cart[1]), pt), 0, rtol=0, atol=ERROR_TOLERANCE
+        np.dot(cross_product, pt_xyz), 0, rtol=MACHINE_EPSILON, atol=MACHINE_EPSILON
     ):
         return False
 
-    if np.isclose(GCRv0_lonlat[0], GCRv1_lonlat[0], rtol=0, atol=ERROR_TOLERANCE):
-        # If the pt and the GCA are on the same longitude (the y coordinates are the same)
-        if np.isclose(GCRv0_lonlat[0], pt_lonlat[0], rtol=0, atol=ERROR_TOLERANCE):
-            # Now use the latitude to determine if the pt falls between the interval
-            return in_between(GCRv0_lonlat[1], pt_lonlat[1], GCRv1_lonlat[1])
-        else:
-            # If the pt and the GCA are not on the same longitude when the GCA is a longnitude arc, then the pt is not on the GCA
-            return False
+    # 3. Check if the point lies within the Great Circle Arc interval
+    pt_a = gca_a_xyz - pt_xyz
+    pt_b = gca_b_xyz - pt_xyz
 
-    # If the longnitude span is exactly 180 degree, then the GCA goes through the pole point
-    if np.isclose(
-        abs(GCRv1_lonlat[0] - GCRv0_lonlat[0]), np.pi, rtol=0, atol=ERROR_TOLERANCE
-    ):
-        # Special case, if the pt is on the pole point, then set its longitude to the GCRv0_lonlat[0]
-        if np.isclose(abs(pt_lonlat[1]), np.pi / 2, rtol=0, atol=ERROR_TOLERANCE):
-            pt_lonlat[0] = GCRv0_lonlat[0]
-        if not np.isclose(
-            GCRv0_lonlat[0], pt_lonlat[0], rtol=0, atol=ERROR_TOLERANCE
-        ) and not np.isclose(
-            GCRv1_lonlat[0], pt_lonlat[0], rtol=0, atol=ERROR_TOLERANCE
-        ):
-            return False
-        else:
-            # Determine the pole latitude and latitude extension
-            if (GCRv0_lonlat[1] > 0 and GCRv1_lonlat[1] > 0) or (
-                GCRv0_lonlat[1] < 0 and GCRv1_lonlat[1] < 0
-            ):
-                pole_lat = np.pi / 2 if GCRv0_lonlat[1] > 0 else -np.pi / 2
-                lat_extend = (
-                    abs(np.pi / 2 - abs(GCRv0_lonlat[1]))
-                    + np.pi / 2
-                    + abs(GCRv1_lonlat[1])
-                )
-            else:
-                pole_lat = _decide_pole_latitude(GCRv0_lonlat[1], GCRv1_lonlat[1])
-                lat_extend = (
-                    abs(np.pi / 2 - abs(GCRv0_lonlat[1]))
-                    + np.pi / 2
-                    + abs(GCRv1_lonlat[1])
-                )
+    # Use the dot product to determine the sign of the angle between pt_a and pt_b
+    cos_theta = np.dot(pt_a, pt_b)
 
-            if is_directed and lat_extend >= np.pi:
-                raise ValueError(
-                    "The input GCA spans more than 180 degrees. Please divide it into two."
-                )
-
-            return in_between(GCRv0_lonlat[1], pt_lonlat[1], pole_lat) or in_between(
-                pole_lat, pt_lonlat[1], GCRv1_lonlat[1]
-            )
-
-    if is_directed:
-        # The anti-meridian case Sufficient condition: absolute difference between the longitudes of the two
-        # vertices is greater than 180 degrees (π radians): abs(GCRv1_lon - GCRv0_lon) > π
-        if abs(GCRv1_lonlat[0] - GCRv0_lonlat[0]) > np.pi:
-            # The necessary condition: the pt longitude is on the opposite side of the anti-meridian
-            # Case 1: where 0 --> x0--> 180 -->x1 -->0 case is lager than the 180degrees (pi radians)
-            if GCRv0_lonlat[0] <= np.pi <= GCRv1_lonlat[0]:
-                raise ValueError(
-                    "The input Great Circle Arc span is larger than 180 degree, please break it into two."
-                )
-
-            # The necessary condition: the pt longitude is on the opposite side of the anti-meridian
-            # Case 2: The anti-meridian case where 180 -->x0 --> 0 lon --> x1 --> 180 lon
-            elif 2 * np.pi > GCRv0_lonlat[0] > np.pi > GCRv1_lonlat[0] > 0:
-                return in_between(
-                    GCRv0_lonlat[0], pt_lonlat[0], 2 * np.pi
-                ) or in_between(0, pt_lonlat[0], GCRv1_lonlat[0])
-
-        # The non-anti-meridian case.
-        else:
-            return in_between(GCRv0_lonlat[0], pt_lonlat[0], GCRv1_lonlat[0])
+    # Return True if the point lies within the interval (smaller arc)
+    if cos_theta < 0:
+        return True
+    elif isclose(cos_theta, 0.0, atol=MACHINE_EPSILON):
+        # set error tolerance to 0.0
+        return True
     else:
-        # The undirected case
-        # sort the longitude
-        GCRv0_lonlat_min, GCRv1_lonlat_max = sorted([GCRv0_lonlat[0], GCRv1_lonlat[0]])
-        if np.pi > GCRv1_lonlat_max - GCRv0_lonlat_min >= 0.0:
-            return in_between(GCRv0_lonlat[0], pt_lonlat[0], GCRv1_lonlat[0])
-        else:
-            return in_between(GCRv1_lonlat_max, pt_lonlat[0], 2 * np.pi) or in_between(
-                0.0, pt_lonlat[0], GCRv0_lonlat_min
-            )
+        return False
 
 
+@njit(cache=True)
 def in_between(p, q, r) -> bool:
     """Determines whether the number q is between p and r.
 
@@ -190,6 +112,7 @@ def in_between(p, q, r) -> bool:
     return p <= q <= r or r <= q <= p
 
 
+@njit(cache=True)
 def _decide_pole_latitude(lat1, lat2):
     """Determine the pole latitude based on the latitudes of two points on a
     Great Circle Arc (GCA).
@@ -229,38 +152,49 @@ def _decide_pole_latitude(lat1, lat2):
     return closest_pole
 
 
-def _angle_of_2_vectors(u, v):
-    """Calculate the angle between two 3D vectors u and v in radians. Can be
-    used to calcualte the span of a GCR.
+@njit(cache=True)
+def max3(a, b, c):
+    if a >= b and a >= c:
+        return a
+    elif b >= c:
+        return b
+    else:
+        return c
 
-    Parameters
-    ----------
-    u : numpy.ndarray (float)
-        The first 3D vector.
-    v : numpy.ndarray (float)
-        The second 3D vector.
 
-    Returns
-    -------
-    float
-        The angle between u and v in radians.
+@njit(cache=True)
+def min3(a, b, c):
+    if a <= b and a <= c:
+        return a
+    elif b <= c:
+        return b
+    else:
+        return c
+
+
+@njit(cache=True)
+def clip_scalar(a, a_min, a_max):
+    if a < a_min:
+        return a_min
+    elif a > a_max:
+        return a_max
+    else:
+        return a
+
+
+@njit(cache=True)
+def extreme_gca_latitude(gca_cart, gca_lonlat, extreme_type):
     """
-    v_norm_times_u = np.linalg.norm(v) * u
-    u_norm_times_v = np.linalg.norm(u) * v
-    vec_minus = v_norm_times_u - u_norm_times_v
-    vec_sum = v_norm_times_u + u_norm_times_v
-    angle_u_v_rad = 2 * np.arctan2(np.linalg.norm(vec_minus), np.linalg.norm(vec_sum))
-    return angle_u_v_rad
-
-
-def extreme_gca_latitude(gca_cart, extreme_type):
-    """Calculate the maximum or minimum latitude of a great circle arc defined
+    Calculate the maximum or minimum latitude of a great circle arc defined
     by two 3D points.
 
     Parameters
     ----------
     gca_cart : numpy.ndarray
         An array containing two 3D vectors that define a great circle arc.
+
+    gca_lonlat : numpy.ndarray
+        An array containing the longitude and latitude of the two points.
 
     extreme_type : str
         The type of extreme latitude to calculate. Must be either 'max' or 'min'.
@@ -275,34 +209,163 @@ def extreme_gca_latitude(gca_cart, extreme_type):
     ValueError
         If `extreme_type` is not 'max' or 'min'.
     """
-    extreme_type = extreme_type.lower()
-
-    if extreme_type not in ("max", "min"):
+    # Validate extreme_type
+    if (extreme_type != "max") and (extreme_type != "min"):
         raise ValueError("extreme_type must be either 'max' or 'min'")
 
-    n1, n2 = gca_cart
-    dot_n1_n2 = np.dot(n1, n2)
+    # Extract the two points
+    n1 = gca_cart[0]
+    n2 = gca_cart[1]
+
+    # Compute dot product
+    dot_n1_n2 = dot(n1, n2)
+
+    # Compute denominator
     denom = (n1[2] + n2[2]) * (dot_n1_n2 - 1.0)
-    d_a_max = (n1[2] * dot_n1_n2 - n2[2]) / denom
 
-    d_a_max = (
-        np.clip(d_a_max, 0, 1)
-        if np.isclose(d_a_max, [0, 1], atol=ERROR_TOLERANCE).any()
-        else d_a_max
-    )
+    # Initialize latitudes
+    lon_n1, lat_n1 = gca_lonlat[0]
+    lon_n2, lat_n2 = gca_lonlat[1]
 
-    _, lat_n1 = _xyz_to_lonlat_rad(n1[0], n1[1], n1[2])
-    _, lat_n2 = _xyz_to_lonlat_rad(n2[0], n2[1], n2[2])
+    # Check if denominator is zero to avoid division by zero
+    if denom != 0.0:
+        d_a_max = (n1[2] * dot_n1_n2 - n2[2]) / denom
 
-    if 0 < d_a_max < 1:
-        node3 = (1 - d_a_max) * n1 + d_a_max * n2
-        node3 = np.array(_normalize_xyz(node3[0], node3[1], node3[2]))
-        d_lat_rad = np.arcsin(np.clip(node3[2], -1, 1))
+        # Handle cases where d_a_max is very close to 0 or 1
+        if isclose(d_a_max, 0.0, atol=ERROR_TOLERANCE) or isclose(
+            d_a_max, 1.0, atol=ERROR_TOLERANCE
+        ):
+            d_a_max = clip_scalar(d_a_max, 0.0, 1.0)
 
-        return (
-            max(d_lat_rad, lat_n1, lat_n2)
-            if extreme_type == "max"
-            else min(d_lat_rad, lat_n1, lat_n2)
-        )
+        # Check if d_a_max is within the valid range
+        if (d_a_max > 0.0) and (d_a_max < 1.0):
+            # Compute the intermediate point on the GCA
+            node3 = (1.0 - d_a_max) * n1 + d_a_max * n2
+
+            # Normalize the intermediate point
+            x, y, z = _normalize_xyz_scalar(node3[0], node3[1], node3[2])
+            node3_normalized = np.empty(3)
+            node3_normalized[0] = x
+            node3_normalized[1] = y
+            node3_normalized[2] = z
+
+            # Compute latitude of the intermediate point
+            d_lat_rad = math.asin(clip_scalar(node3_normalized[2], -1.0, 1.0))
+
+            # Return the extreme latitude
+            if extreme_type == "max":
+                return max3(d_lat_rad, lat_n1, lat_n2)
+            else:
+                return min3(d_lat_rad, lat_n1, lat_n2)
+
+    # If denom is zero or d_a_max is not in (0,1), return max or min of lat_n1 and lat_n2
+    if extreme_type == "max":
+        return max(lat_n1, lat_n2)
     else:
-        return max(lat_n1, lat_n2) if extreme_type == "max" else min(lat_n1, lat_n2)
+        return min(lat_n1, lat_n2)
+
+
+@njit(cache=True)
+def extreme_gca_z(gca_cart, extreme_type):
+    """
+    Calculate the maximum or minimum latitude of a great circle arc defined
+    by two 3D points.
+
+    Parameters
+    ----------
+    gca_cart : numpy.ndarray
+        An array containing two 3D vectors that define a great circle arc.
+
+    extreme_type : str
+        The type of extreme latitude to calculate. Must be either 'max' or 'min'.
+
+    Returns
+    -------
+    float
+        The maximum or minimum z of the great circle arc
+
+    Raises
+    ------
+    ValueError
+        If `extreme_type` is not 'max' or 'min'.
+    """
+
+    # Validate extreme_type
+    if (extreme_type != "max") and (extreme_type != "min"):
+        raise ValueError("extreme_type must be either 'max' or 'min'")
+
+    # Extract the two points
+    n1 = gca_cart[0]
+    n2 = gca_cart[1]
+
+    # Compute dot product
+    dot_n1_n2 = dot(n1, n2)
+
+    # Compute denominator
+    denom = (n1[2] + n2[2]) * (dot_n1_n2 - 1.0)
+
+    # (z) coordinate
+    z_n1 = gca_cart[0][2]
+    z_n2 = gca_cart[1][2]
+
+    # Check if the denominator is zero to avoid division by zero
+    if denom != 0.0:
+        d_a_max = (n1[2] * dot_n1_n2 - n2[2]) / denom
+
+        # Handle cases where d_a_max is very close to 0 or 1
+        if isclose(d_a_max, 0.0, atol=ERROR_TOLERANCE) or isclose(
+            d_a_max, 1.0, atol=ERROR_TOLERANCE
+        ):
+            d_a_max = clip_scalar(d_a_max, 0.0, 1.0)
+
+        # Check if d_a_max is within the valid range
+        if (d_a_max > 0.0) and (d_a_max < 1.0):
+            # Compute the intermediate point on the GCA
+            node3 = (1.0 - d_a_max) * n1 + d_a_max * n2
+
+            # Normalize the intermediate point
+            x, y, z = _normalize_xyz_scalar(node3[0], node3[1], node3[2])
+            node3_normalized = np.empty(3)
+            node3_normalized[0] = x
+            node3_normalized[1] = y
+            node3_normalized[2] = z
+
+            d_z = clip_scalar(node3_normalized[2], -1.0, 1.0)
+
+            if extreme_type == "max":
+                return max3(d_z, z_n1, z_n2)
+            else:
+                return min3(d_z, z_n1, z_n2)
+
+    # If denom is zero or d_a_max is not in (0,1), return max or min of lat_n1 and lat_n2
+    if extreme_type == "max":
+        return max(z_n1, z_n2)
+    else:
+        return min(z_n1, z_n2)
+
+
+@njit(cache=True)
+def compute_arc_length(pt_a, pt_b):
+    """
+    Compute the great circle arc length between two points on a unit sphere at constant latitude.
+
+    Parameters
+    ----------
+    pt_a : tuple or array-like
+        First point coordinates (x, y, z) on unit sphere
+    pt_b : tuple or array-like
+        Second point coordinates (x, y, z) on unit sphere
+
+    Returns
+    -------
+    float
+        Arc length between the points at their constant latitude
+    """
+    x1, y1, z1 = pt_a
+    x2, y2, z2 = pt_b
+    rho = np.sqrt(1.0 - z1 * z2)
+    cross_2d = x1 * y2 - y1 * x2
+    dot_2d = x1 * x2 + y1 * y2
+    delta_theta = np.atan2(cross_2d, dot_2d)
+
+    return rho * abs(delta_theta)
