@@ -2,15 +2,19 @@ import numpy as np
 
 from uxarray.grid.coordinates import _lonlat_rad_to_xyz
 
-from numba import njit, config
-from uxarray.constants import ENABLE_JIT_CACHE, ENABLE_JIT
-
-config.DISABLE_JIT = not ENABLE_JIT
+from numba import njit
+from uxarray.constants import ERROR_TOLERANCE
 
 
-@njit(cache=ENABLE_JIT_CACHE)
+@njit(cache=True)
 def calculate_face_area(
-    x, y, z, quadrature_rule="gaussian", order=4, coords_type="spherical"
+    x,
+    y,
+    z,
+    quadrature_rule="gaussian",
+    order=4,
+    coords_type="spherical",
+    latitude_adjusted_area=False,
 ):
     """Calculate area of a face on sphere.
 
@@ -38,6 +42,9 @@ def calculate_face_area(
     coords_type : str, optional
         coordinate type, default is spherical, can be cartesian also.
 
+    latitude_adjusted_area : bool, optional
+        If True, performs the check if any face consists of an edge that has constant latitude, modifies the area of that face by applying the correction term due to that edge. Default is False.
+
     Returns
     -------
     area : double
@@ -48,9 +55,9 @@ def calculate_face_area(
     order = order
 
     if quadrature_rule == "gaussian":
-        dG, dW = get_gauss_quadratureDG(order)
+        dG, dW = get_gauss_quadrature_dg(order)
     elif quadrature_rule == "triangular":
-        dG, dW = get_tri_quadratureDG(order)
+        dG, dW = get_tri_quadrature_dg(order)
     else:
         raise ValueError("Invalid quadrature rule, specify gaussian or triangular")
 
@@ -59,22 +66,30 @@ def calculate_face_area(
     # num triangles is two less than the total number of nodes
     num_triangles = num_nodes - 2
 
+    # TODO: Remove/clarify option for spherical coordinates
+    if coords_type == "spherical":
+        # Preallocate arrays for Cartesian coordinates
+        n_points = len(x)
+        x_cartesian = np.empty(n_points)
+        y_cartesian = np.empty(n_points)
+        z_cartesian = np.empty(n_points)
+
+        # Convert all points to Cartesian coordinates using an explicit loop
+        for i in range(n_points):
+            lon_rad = np.deg2rad(x[i])
+            lat_rad = np.deg2rad(y[i])
+            cartesian = _lonlat_rad_to_xyz(lon_rad, lat_rad)
+            x_cartesian[i], y_cartesian[i], z_cartesian[i] = cartesian
+
+        x, y, z = x_cartesian, y_cartesian, z_cartesian
+
     # Using tempestremap GridElements: https://github.com/ClimateGlobalChange/tempestremap/blob/master/src/GridElements.cpp
     # loop through all sub-triangles of face
+    total_correction = 0.0
     for j in range(0, num_triangles):
-        node1 = np.array([x[0], y[0], z[0]], dtype=np.float64)
-        node2 = np.array([x[j + 1], y[j + 1], z[j + 1]], dtype=np.float64)
-        node3 = np.array([x[j + 2], y[j + 2], z[j + 2]], dtype=np.float64)
-
-        if coords_type == "spherical":
-            node1 = _lonlat_rad_to_xyz(np.deg2rad(x[0]), np.deg2rad(y[0]))
-            node1 = np.asarray(node1)
-
-            node2 = _lonlat_rad_to_xyz(np.deg2rad(x[j + 1]), np.deg2rad(y[j + 1]))
-            node2 = np.asarray(node2)
-
-            node3 = _lonlat_rad_to_xyz(np.deg2rad(x[j + 2]), np.deg2rad(y[j + 2]))
-            node3 = np.asarray(node3)
+        node1 = np.array([x[0], y[0], z[0]], dtype=x.dtype)
+        node2 = np.array([x[j + 1], y[j + 1], z[j + 1]], dtype=x.dtype)
+        node3 = np.array([x[j + 2], y[j + 2], z[j + 2]], dtype=x.dtype)
 
         for p in range(len(dW)):
             if quadrature_rule == "gaussian":
@@ -95,10 +110,89 @@ def calculate_face_area(
                 area += dW[p] * jacobian
                 jacobian += jacobian
 
+    # check if the any edge is on the line of constant latitude
+    # which means we need to check edges for same z-coordinates and call area correction routine
+    correction = 0.0
+    # TODO: Make this work when latitude_adjusted_area is False and each edge has a flag that indicates if it is a constant latitude edge
+    if latitude_adjusted_area:
+        for i in range(num_nodes):
+            node1 = np.array([x[i], y[i], z[i]], dtype=x.dtype)
+            node2 = np.array(
+                [
+                    x[(i + 1) % num_nodes],
+                    y[(i + 1) % num_nodes],
+                    z[(i + 1) % num_nodes],
+                ],
+                dtype=x.dtype,
+            )
+            # Check if z-coordinates are approximately equal
+            if np.isclose(node1[2], node2[2], atol=ERROR_TOLERANCE):
+                if node1[2] == 0:
+                    # check if z-coordinates are 0 - Equator
+                    continue
+
+                # Check if the edge passes through a pole
+                passes_through_pole = edge_passes_through_pole(node1, node2)
+                if passes_through_pole:
+                    # Skip the edge if it passes through a pole
+                    continue
+
+                z_sign = np.sign(node1[2])
+                # Convert Cartesian coordinates to longitude
+                lon1 = np.arctan2(node1[1], node1[0])
+                lon2 = np.arctan2(node2[1], node2[0])
+                # Calculate the longitude difference in radians
+                lon_diff = lon2 - lon1
+                # Skip the edge if it spans more than 180 degrees of longitude
+                if abs(lon_diff) > np.pi:
+                    continue
+                else:
+                    # Calculate the correction term
+                    correction = area_correction(node1, node2)
+
+                # Check if the longitude is increasing in the northern hemisphere or decreasing in the southern hemisphere
+                if (z_sign > 0 and lon_diff > 0) or (z_sign < 0 and lon_diff < 0):
+                    correction = -correction
+
+                total_correction += correction
+
+    if total_correction != 0.0:
+        area += total_correction
+
     return area, jacobian
 
 
-@njit(cache=ENABLE_JIT_CACHE)
+@njit(cache=True)
+def edge_passes_through_pole(node1, node2):
+    """
+    Check if the edge passes through a pole.
+
+    Parameters:
+    - node1: first node of the edge (normalized).
+    - node2: second node of the edge (normalized).
+
+    Returns:
+    - bool: True if the edge passes through a pole, False otherwise.
+    """
+    # Calculate the normal vector to the plane defined by the origin, node1, and node2
+    n = np.cross(node1, node2)
+
+    # Check for numerical stability issues with the normal vector
+    if np.allclose(n, 0):
+        # Handle cases where the cross product is near zero, such as when nodes are nearly identical or opposite
+        return False
+
+    # North and South Pole vectors
+    p_north = np.array([0.0, 0.0, 1.0])
+    p_south = np.array([0.0, 0.0, -1.0])
+
+    # Check if the normal vector is orthogonal to either pole
+    return np.isclose(np.dot(n, p_north), 0, atol=ERROR_TOLERANCE) or np.isclose(
+        np.dot(n, p_south), 0, atol=ERROR_TOLERANCE
+    )
+
+
+@njit(cache=True)
 def get_all_face_area_from_coords(
     x,
     y,
@@ -109,6 +203,7 @@ def get_all_face_area_from_coords(
     quadrature_rule="triangular",
     order=4,
     coords_type="spherical",
+    latitude_adjusted_area=False,
 ):
     """Given coords, connectivity and other area calculation params, this
     routine loop over all faces and return an numpy array with areas of each
@@ -140,10 +235,17 @@ def get_all_face_area_from_coords(
     coords_type : str, optional
         coordinate type, default is spherical, can be cartesian also.
 
+    latitude_adjusted_area : bool, optional
+        If True, performs the check if any face consists of an edge that has constant latitude, modifies the area of that face by applying the correction term due to that edge. Default is False.
+
     Returns
     -------
-    area of all faces : ndarray
+    area and jacobian of all faces : ndarray, ndarray
     """
+    # this casting helps to prevent the type mismatch
+    x = np.asarray(x, dtype=np.float64)
+    y = np.asarray(y, dtype=np.float64)
+    z = np.asarray(z, dtype=np.float64)
 
     n_face, n_max_face_nodes = face_nodes.shape
 
@@ -164,7 +266,13 @@ def get_all_face_area_from_coords(
 
         # After getting all the nodes of a face assembled call the  cal. face area routine
         face_area, face_jacobian = calculate_face_area(
-            face_x, face_y, face_z, quadrature_rule, order, coords_type
+            face_x,
+            face_y,
+            face_z,
+            quadrature_rule,
+            order,
+            coords_type,
+            latitude_adjusted_area,
         )
         # store current face area
         area[face_idx] = face_area
@@ -173,8 +281,39 @@ def get_all_face_area_from_coords(
     return area, jacobian
 
 
-@njit(cache=ENABLE_JIT_CACHE)
-def calculate_spherical_triangle_jacobian(node1, node2, node3, dA, dB):
+@njit(cache=True)
+def area_correction(node1, node2):
+    """
+    Calculate the area correction A using the given formula.
+
+    Parameters:
+    - node1: first node of the edge (normalized).
+    - node2: second node of the edge (normalized).
+    - z: z-coordinate (shared by both points and part of the formula, normalized).
+
+    Returns:
+    - A: correction term of the area, when one of the edges is a line of constant latitude
+    """
+    x1, y1, z = node1
+    x2, y2, _ = node2
+
+    # Calculate terms
+    term1 = x1 * y2 - x2 * y1
+    den2 = x1 * x2 + y1 * y2
+    den1 = x1**2 + y1**2 + den2
+
+    # Compute angles using arctan2
+    angle1 = np.arctan2(z * term1, den1)
+    angle2 = np.arctan2(term1, den2)
+
+    # Compute A
+    A = np.abs(2 * angle1 - z * angle2)
+
+    return A
+
+
+@njit(cache=True)
+def calculate_spherical_triangle_jacobian(node1, node2, node3, d_a, d_b):
     """Calculate Jacobian of a spherical triangle. This is a helper function
     for calculating face area.
 
@@ -189,82 +328,82 @@ def calculate_spherical_triangle_jacobian(node1, node2, node3, dA, dB):
     node3 : list, required
         Third node of the triangle
 
-    dA : float, required
+    d_a : float, required
         quadrature point
 
-    dB : float, required
+    d_b : float, required
         quadrature point
 
     Returns
     -------
     jacobian : float
     """
-    dF = np.array(
+    d_f = np.array(
         [
-            (1.0 - dB) * ((1.0 - dA) * node1[0] + dA * node2[0]) + dB * node3[0],
-            (1.0 - dB) * ((1.0 - dA) * node1[1] + dA * node2[1]) + dB * node3[1],
-            (1.0 - dB) * ((1.0 - dA) * node1[2] + dA * node2[2]) + dB * node3[2],
+            (1.0 - d_b) * ((1.0 - d_a) * node1[0] + d_a * node2[0]) + d_b * node3[0],
+            (1.0 - d_b) * ((1.0 - d_a) * node1[1] + d_a * node2[1]) + d_b * node3[1],
+            (1.0 - d_b) * ((1.0 - d_a) * node1[2] + d_a * node2[2]) + d_b * node3[2],
         ]
     )
 
-    dDaF = np.array(
+    d_da_f = np.array(
         [
-            (1.0 - dB) * (node2[0] - node1[0]),
-            (1.0 - dB) * (node2[1] - node1[1]),
-            (1.0 - dB) * (node2[2] - node1[2]),
+            (1.0 - d_b) * (node2[0] - node1[0]),
+            (1.0 - d_b) * (node2[1] - node1[1]),
+            (1.0 - d_b) * (node2[2] - node1[2]),
         ]
     )
 
-    dDbF = np.array(
+    d_db_f = np.array(
         [
-            -(1.0 - dA) * node1[0] - dA * node2[0] + node3[0],
-            -(1.0 - dA) * node1[1] - dA * node2[1] + node3[1],
-            -(1.0 - dA) * node1[2] - dA * node2[2] + node3[2],
+            -(1.0 - d_a) * node1[0] - d_a * node2[0] + node3[0],
+            -(1.0 - d_a) * node1[1] - d_a * node2[1] + node3[1],
+            -(1.0 - d_a) * node1[2] - d_a * node2[2] + node3[2],
         ]
     )
 
-    dInvR = 1.0 / np.sqrt(dF[0] * dF[0] + dF[1] * dF[1] + dF[2] * dF[2])
+    d_inv_r = 1.0 / np.sqrt(d_f[0] * d_f[0] + d_f[1] * d_f[1] + d_f[2] * d_f[2])
 
-    dDaG = np.array(
+    d_da_g = np.array(
         [
-            dDaF[0] * (dF[1] * dF[1] + dF[2] * dF[2])
-            - dF[0] * (dDaF[1] * dF[1] + dDaF[2] * dF[2]),
-            dDaF[1] * (dF[0] * dF[0] + dF[2] * dF[2])
-            - dF[1] * (dDaF[0] * dF[0] + dDaF[2] * dF[2]),
-            dDaF[2] * (dF[0] * dF[0] + dF[1] * dF[1])
-            - dF[2] * (dDaF[0] * dF[0] + dDaF[1] * dF[1]),
+            d_da_f[0] * (d_f[1] * d_f[1] + d_f[2] * d_f[2])
+            - d_f[0] * (d_da_f[1] * d_f[1] + d_da_f[2] * d_f[2]),
+            d_da_f[1] * (d_f[0] * d_f[0] + d_f[2] * d_f[2])
+            - d_f[1] * (d_da_f[0] * d_f[0] + d_da_f[2] * d_f[2]),
+            d_da_f[2] * (d_f[0] * d_f[0] + d_f[1] * d_f[1])
+            - d_f[2] * (d_da_f[0] * d_f[0] + d_da_f[1] * d_f[1]),
         ]
     )
 
-    dDbG = np.array(
+    d_db_g = np.array(
         [
-            dDbF[0] * (dF[1] * dF[1] + dF[2] * dF[2])
-            - dF[0] * (dDbF[1] * dF[1] + dDbF[2] * dF[2]),
-            dDbF[1] * (dF[0] * dF[0] + dF[2] * dF[2])
-            - dF[1] * (dDbF[0] * dF[0] + dDbF[2] * dF[2]),
-            dDbF[2] * (dF[0] * dF[0] + dF[1] * dF[1])
-            - dF[2] * (dDbF[0] * dF[0] + dDbF[1] * dF[1]),
+            d_db_f[0] * (d_f[1] * d_f[1] + d_f[2] * d_f[2])
+            - d_f[0] * (d_db_f[1] * d_f[1] + d_db_f[2] * d_f[2]),
+            d_db_f[1] * (d_f[0] * d_f[0] + d_f[2] * d_f[2])
+            - d_f[1] * (d_db_f[0] * d_f[0] + d_db_f[2] * d_f[2]),
+            d_db_f[2] * (d_f[0] * d_f[0] + d_f[1] * d_f[1])
+            - d_f[2] * (d_db_f[0] * d_f[0] + d_db_f[1] * d_f[1]),
         ]
     )
 
-    dDenomTerm = dInvR * dInvR * dInvR
+    d_denom_term = d_inv_r * d_inv_r * d_inv_r
 
-    dDaG *= dDenomTerm
-    dDbG *= dDenomTerm
+    d_da_g *= d_denom_term
+    d_db_g *= d_denom_term
 
     #  Cross product gives local Jacobian
-    nodeCross = np.cross(dDaG, dDbG)
-    dJacobian = np.sqrt(
-        nodeCross[0] * nodeCross[0]
-        + nodeCross[1] * nodeCross[1]
-        + nodeCross[2] * nodeCross[2]
+    node_cross = np.cross(d_da_g, d_db_g)
+    d_jacobian = np.sqrt(
+        node_cross[0] * node_cross[0]
+        + node_cross[1] * node_cross[1]
+        + node_cross[2] * node_cross[2]
     )
 
-    return dJacobian
+    return d_jacobian
 
 
-@njit(cache=ENABLE_JIT_CACHE)
-def calculate_spherical_triangle_jacobian_barycentric(node1, node2, node3, dA, dB):
+@njit(cache=True)
+def calculate_spherical_triangle_jacobian_barycentric(node1, node2, node3, d_a, d_b):
     """Calculate Jacobian of a spherical triangle. This is a helper function
     for calculating face area.
 
@@ -279,119 +418,123 @@ def calculate_spherical_triangle_jacobian_barycentric(node1, node2, node3, dA, d
     node3 : list, required
         Third node of the triangle
 
-    dA : float, required
+    d_a : float, required
         first component of barycentric coordinates of quadrature point
 
-    dB : float, required
+    d_b : float, required
         second component of barycentric coordinates of quadrature point
 
     Returns
     -------
     jacobian : float
     """
-
-    dF = np.array(
+    # Calculate the position vector d_f
+    d_f = np.array(
         [
-            dA * node1[0] + dB * node2[0] + (1.0 - dA - dB) * node3[0],
-            dA * node1[1] + dB * node2[1] + (1.0 - dA - dB) * node3[1],
-            dA * node1[2] + dB * node2[2] + (1.0 - dA - dB) * node3[2],
+            d_a * node1[0] + d_b * node2[0] + (1.0 - d_a - d_b) * node3[0],
+            d_a * node1[1] + d_b * node2[1] + (1.0 - d_a - d_b) * node3[1],
+            d_a * node1[2] + d_b * node2[2] + (1.0 - d_a - d_b) * node3[2],
         ]
     )
 
-    dDaF = np.array([node1[0] - node3[0], node1[1] - node3[1], node1[2] - node3[2]])
+    # Calculate the gradients d_da_f and d_db_f
+    d_da_f = np.array([node1[0] - node3[0], node1[1] - node3[1], node1[2] - node3[2]])
+    d_db_f = np.array([node2[0] - node3[0], node2[1] - node3[1], node2[2] - node3[2]])
 
-    dDbF = np.array([node2[0] - node3[0], node2[1] - node3[1], node2[2] - node3[2]])
+    # Calculate the inverse radius
+    d_inv_r = 1.0 / np.sqrt(d_f[0] * d_f[0] + d_f[1] * d_f[1] + d_f[2] * d_f[2])
 
-    dInvR = 1.0 / np.sqrt(dF[0] * dF[0] + dF[1] * dF[1] + dF[2] * dF[2])
-
-    dDaG = np.array(
+    # Calculate the gradients d_da_g and d_db_g
+    d_da_g = np.array(
         [
-            dDaF[0] * (dF[1] * dF[1] + dF[2] * dF[2])
-            - dF[0] * (dDaF[1] * dF[1] + dDaF[2] * dF[2]),
-            dDaF[1] * (dF[0] * dF[0] + dF[2] * dF[2])
-            - dF[1] * (dDaF[0] * dF[0] + dDaF[2] * dF[2]),
-            dDaF[2] * (dF[0] * dF[0] + dF[1] * dF[1])
-            - dF[2] * (dDaF[0] * dF[0] + dDaF[1] * dF[1]),
+            d_da_f[0] * (d_f[1] * d_f[1] + d_f[2] * d_f[2])
+            - d_f[0] * (d_da_f[1] * d_f[1] + d_da_f[2] * d_f[2]),
+            d_da_f[1] * (d_f[0] * d_f[0] + d_f[2] * d_f[2])
+            - d_f[1] * (d_da_f[0] * d_f[0] + d_da_f[2] * d_f[2]),
+            d_da_f[2] * (d_f[0] * d_f[0] + d_f[1] * d_f[1])
+            - d_f[2] * (d_da_f[0] * d_f[0] + d_da_f[1] * d_f[1]),
         ]
     )
 
-    dDbG = np.array(
+    d_db_g = np.array(
         [
-            dDbF[0] * (dF[1] * dF[1] + dF[2] * dF[2])
-            - dF[0] * (dDbF[1] * dF[1] + dDbF[2] * dF[2]),
-            dDbF[1] * (dF[0] * dF[0] + dF[2] * dF[2])
-            - dF[1] * (dDbF[0] * dF[0] + dDbF[2] * dF[2]),
-            dDbF[2] * (dF[0] * dF[0] + dF[1] * dF[1])
-            - dF[2] * (dDbF[0] * dF[0] + dDbF[1] * dF[1]),
+            d_db_f[0] * (d_f[1] * d_f[1] + d_f[2] * d_f[2])
+            - d_f[0] * (d_db_f[1] * d_f[1] + d_db_f[2] * d_f[2]),
+            d_db_f[1] * (d_f[0] * d_f[0] + d_f[2] * d_f[2])
+            - d_f[1] * (d_db_f[0] * d_f[0] + d_db_f[2] * d_f[2]),
+            d_db_f[2] * (d_f[0] * d_f[0] + d_f[1] * d_f[1])
+            - d_f[2] * (d_db_f[0] * d_f[0] + d_db_f[1] * d_f[1]),
         ]
     )
 
-    dDenomTerm = dInvR * dInvR * dInvR
+    # Calculate the denominator term
+    d_denom_term = d_inv_r * d_inv_r * d_inv_r
 
-    dDaG *= dDenomTerm
-    dDbG *= dDenomTerm
+    # Scale the gradients
+    d_da_g *= d_denom_term
+    d_db_g *= d_denom_term
 
-    #  Cross product gives local Jacobian
-    nodeCross = np.cross(dDaG, dDbG)
-    dJacobian = np.sqrt(
-        nodeCross[0] * nodeCross[0]
-        + nodeCross[1] * nodeCross[1]
-        + nodeCross[2] * nodeCross[2]
+    # Calculate the cross product
+    node_cross = np.cross(d_da_g, d_db_g)
+
+    # Calculate the Jacobian
+    d_jacobian = np.sqrt(
+        node_cross[0] * node_cross[0]
+        + node_cross[1] * node_cross[1]
+        + node_cross[2] * node_cross[2]
     )
 
-    return 0.5 * dJacobian
+    return 0.5 * d_jacobian
 
 
-@njit(cache=ENABLE_JIT_CACHE)
-def get_gauss_quadratureDG(nCount):
+@njit(cache=True)
+def get_gauss_quadrature_dg(n_count):
     """Gauss Quadrature Points for integration.
 
     Parameters
     ----------
-    nCount : int, required
+    n_count : int, required
          Degree of quadrature points required, supports: 1 to 10.
 
     Returns
     -------
-        dG : double
-            numpy array of size ncount, quadrature points. Scaled before returning.
-        dW : double
-            numpy array of size ncount x 3, weights. Scaled before returning.
+        d_g : double
+            numpy array of size n_count, quadrature points. Scaled before returning.
+        d_w : double
+            numpy array of size n_count x 3, weights. Scaled before returning.
 
     Raises
     ------
        RuntimeError: Invalid degree
     """
     # Degree 1
-    if nCount == 1:
-        dG = np.array([[0.0]])
-        dW = np.array([+2.0])
+    if n_count == 1:
+        d_g = np.array([[0.0]])
+        d_w = np.array([2.0])
 
     # Degree 2
-    elif nCount == 2:
-        dG = np.array([[-0.5773502691896257, +0.5773502691896257]])
-        dW = np.array([+1.0, +1.0])
+    elif n_count == 2:
+        d_g = np.array([[-0.5773502691896257, 0.5773502691896257]])
+        d_w = np.array([1.0, 1.0])
 
     # Degree 3
-    elif nCount == 3:
-        dG = np.array([[-0.7745966692414834, 0.0, +0.7745966692414834]])
-
-        dW = np.array([+0.5555555555555556, +0.8888888888888888, +0.5555555555555556])
+    elif n_count == 3:
+        d_g = np.array([[-0.7745966692414834, 0.0, 0.7745966692414834]])
+        d_w = np.array([0.5555555555555556, 0.8888888888888888, 0.5555555555555556])
 
     # Degree 4
-    elif nCount == 4:
-        dG = np.array(
+    elif n_count == 4:
+        d_g = np.array(
             [
                 [
                     -0.8611363115940526,
                     -0.3399810435848563,
-                    +0.3399810435848563,
-                    +0.8611363115940526,
+                    0.3399810435848563,
+                    0.8611363115940526,
                 ]
             ]
         )
-
-        dW = np.array(
+        d_w = np.array(
             [
                 0.3478548451374538,
                 0.6521451548625461,
@@ -401,20 +544,19 @@ def get_gauss_quadratureDG(nCount):
         )
 
     # Degree 5
-    elif nCount == 5:
-        dG = np.array(
+    elif n_count == 5:
+        d_g = np.array(
             [
                 [
                     -0.9061798459386640,
                     -0.5384693101056831,
                     0.0,
-                    +0.5384693101056831,
-                    +0.9061798459386640,
+                    0.5384693101056831,
+                    0.9061798459386640,
                 ]
             ]
         )
-
-        dW = np.array(
+        d_w = np.array(
             [
                 0.2369268850561891,
                 0.4786286704993665,
@@ -425,21 +567,20 @@ def get_gauss_quadratureDG(nCount):
         )
 
     # Degree 6
-    elif nCount == 6:
-        dG = np.array(
+    elif n_count == 6:
+        d_g = np.array(
             [
                 [
                     -0.9324695142031521,
                     -0.6612093864662645,
                     -0.2386191860831969,
-                    +0.2386191860831969,
-                    +0.6612093864662645,
-                    +0.9324695142031521,
+                    0.2386191860831969,
+                    0.6612093864662645,
+                    0.9324695142031521,
                 ]
             ]
         )
-
-        dW = np.array(
+        d_w = np.array(
             [
                 0.1713244923791704,
                 0.3607615730481386,
@@ -451,22 +592,21 @@ def get_gauss_quadratureDG(nCount):
         )
 
     # Degree 7
-    elif nCount == 7:
-        dG = np.array(
+    elif n_count == 7:
+        d_g = np.array(
             [
                 [
                     -0.9491079123427585,
                     -0.7415311855993945,
                     -0.4058451513773972,
                     0.0,
-                    +0.4058451513773972,
-                    +0.7415311855993945,
-                    +0.9491079123427585,
+                    0.4058451513773972,
+                    0.7415311855993945,
+                    0.9491079123427585,
                 ]
             ]
         )
-
-        dW = np.array(
+        d_w = np.array(
             [
                 0.1294849661688697,
                 0.2797053914892766,
@@ -479,23 +619,22 @@ def get_gauss_quadratureDG(nCount):
         )
 
     # Degree 8
-    elif nCount == 8:
-        dG = np.array(
+    elif n_count == 8:
+        d_g = np.array(
             [
                 [
                     -0.9602898564975363,
                     -0.7966664774136267,
                     -0.5255324099163290,
                     -0.1834346424956498,
-                    +0.1834346424956498,
-                    +0.5255324099163290,
-                    +0.7966664774136267,
-                    +0.9602898564975363,
+                    0.1834346424956498,
+                    0.5255324099163290,
+                    0.7966664774136267,
+                    0.9602898564975363,
                 ]
             ]
         )
-
-        dW = np.array(
+        d_w = np.array(
             [
                 0.1012285362903763,
                 0.2223810344533745,
@@ -509,8 +648,8 @@ def get_gauss_quadratureDG(nCount):
         )
 
     # Degree 9
-    elif nCount == 9:
-        dG = np.array(
+    elif n_count == 9:
+        d_g = np.array(
             [
                 [
                     -1.0,
@@ -518,15 +657,14 @@ def get_gauss_quadratureDG(nCount):
                     -0.677186279510738,
                     -0.363117463826178,
                     0.0,
-                    +0.363117463826178,
-                    +0.677186279510738,
-                    +0.899757995411460,
-                    +1.0,
+                    0.363117463826178,
+                    0.677186279510738,
+                    0.899757995411460,
+                    1.0,
                 ]
             ]
         )
-
-        dW = np.array(
+        d_w = np.array(
             [
                 0.0277777777777778,
                 0.1654953615608055,
@@ -541,8 +679,8 @@ def get_gauss_quadratureDG(nCount):
         )
 
     # Degree 10
-    elif nCount == 10:
-        dG = np.array(
+    elif n_count == 10:
+        d_g = np.array(
             [
                 [
                     -0.9739065285171717,
@@ -550,16 +688,15 @@ def get_gauss_quadratureDG(nCount):
                     -0.6794095682990244,
                     -0.4333953941292472,
                     -0.1488743389816312,
-                    +0.1488743389816312,
-                    +0.4333953941292472,
-                    +0.6794095682990244,
-                    +0.8650633666889845,
-                    +0.9739065285171717,
+                    0.1488743389816312,
+                    0.4333953941292472,
+                    0.6794095682990244,
+                    0.8650633666889845,
+                    0.9739065285171717,
                 ]
             ]
         )
-
-        dW = np.array(
+        d_w = np.array(
             [
                 0.0666713443086881,
                 0.1494513491505806,
@@ -573,37 +710,34 @@ def get_gauss_quadratureDG(nCount):
                 0.0666713443086881,
             ]
         )
-    # else:
-    #     msg = "quadrature order 1 to 10 is supported: ", nCount, " is invalid\n"
-    #     raise ValueError(msg)
 
     # Scale quadrature points
-    dXi0 = 0.0
-    dXi1 = 1.0
-    for i in range(nCount):
-        dG[0][i] = dXi0 + 0.5 * (dXi1 - dXi0) * (dG[0][i] + 1.0)
-        dW[i] = 0.5 * (dXi1 - dXi0) * dW[i]
+    d_xi0 = 0.0
+    d_xi1 = 1.0
+    for i in range(n_count):
+        d_g[0][i] = d_xi0 + 0.5 * (d_xi1 - d_xi0) * (d_g[0][i] + 1.0)
+        d_w[i] = 0.5 * (d_xi1 - d_xi0) * d_w[i]
 
-    return dG, dW
+    return d_g, d_w
 
 
-@njit(cache=ENABLE_JIT_CACHE)
-def get_tri_quadratureDG(nOrder):
+@njit(cache=True)
+def get_tri_quadrature_dg(n_order):
     """Triangular Quadrature Points for integration.
 
     Parameters
     ----------
-    nOrder : int
+    n_order : int
         Integration order, supports: 12, 10, 8, 4 and 1
 
     Returns
     -------
-        dG, dW : ndarray
+        d_g, d_w : ndarray
             points and weights, with dimension order x 3
     """
     # 12th order quadrature rule (33 points)
-    if nOrder == 12:
-        dG = np.array(
+    if n_order == 12:
+        d_g = np.array(
             [
                 [0.023565220452390, 0.488217389773805, 0.488217389773805],
                 [0.488217389773805, 0.023565220452390, 0.488217389773805],
@@ -641,7 +775,7 @@ def get_tri_quadratureDG(nOrder):
             ]
         )
 
-        dW = np.array(
+        d_w = np.array(
             [
                 0.025731066440455,
                 0.025731066440455,
@@ -680,8 +814,8 @@ def get_tri_quadratureDG(nOrder):
         )
 
     # 10th order quadrature rule (25 points)
-    elif nOrder == 10:
-        dG = np.array(
+    elif n_order == 10:
+        d_g = np.array(
             [
                 [0.333333333333333, 0.333333333333333, 0.333333333333333],
                 [0.028844733232685, 0.485577633383657, 0.485577633383657],
@@ -711,7 +845,7 @@ def get_tri_quadratureDG(nOrder):
             ]
         )
 
-        dW = np.array(
+        d_w = np.array(
             [
                 0.090817990382754,
                 0.036725957756467,
@@ -742,8 +876,8 @@ def get_tri_quadratureDG(nOrder):
         )
 
     # 8th order quadrature rule (16 points)
-    elif nOrder == 8:
-        dG = np.array(
+    elif n_order == 8:
+        d_g = np.array(
             [
                 [0.333333333333333, 0.333333333333333, 0.333333333333333],
                 [0.081414823414554, 0.459292588292723, 0.459292588292723],
@@ -764,7 +898,7 @@ def get_tri_quadratureDG(nOrder):
             ]
         )
 
-        dW = np.array(
+        d_w = np.array(
             [
                 0.144315607677787,
                 0.095091634267285,
@@ -786,8 +920,8 @@ def get_tri_quadratureDG(nOrder):
         )
 
     # 4th order quadrature rule (6 points)
-    elif nOrder == 4:
-        dG = np.array(
+    elif n_order == 4:
+        d_g = np.array(
             [
                 [0.108103018168070, 0.445948490915965, 0.445948490915965],
                 [0.445948490915965, 0.108103018168070, 0.445948490915965],
@@ -798,7 +932,7 @@ def get_tri_quadratureDG(nOrder):
             ]
         )
 
-        dW = np.array(
+        d_w = np.array(
             [
                 0.223381589678011,
                 0.223381589678011,
@@ -809,8 +943,8 @@ def get_tri_quadratureDG(nOrder):
             ]
         )
     # 1st order quadrature rule (1 point)
-    elif nOrder == 1:
-        dG = np.array([[0.333333333333333, 0.333333333333333, 0.333333333333333]])
-        dW = np.array([1.000000000000000])
+    elif n_order == 1:
+        d_g = np.array([[0.333333333333333, 0.333333333333333, 0.333333333333333]])
+        d_w = np.array([1.000000000000000])
 
-    return dG, dW
+    return d_g, d_w
