@@ -1,311 +1,322 @@
 import numpy as np
-from uxarray.constants import ERROR_TOLERANCE, INT_FILL_VALUE
-import warnings
-import uxarray.utils.computing as ac_utils
+from uxarray.constants import INT_FILL_VALUE
+
+from numba import njit
 
 
-def _replace_fill_values(grid_var, original_fill, new_fill, new_dtype=None):
-    """Replaces all instances of the current fill value (``original_fill``) in
-    (``grid_var``) with (``new_fill``) and converts to the dtype defined by
-    (``new_dtype``)
-
-    Parameters
-    ----------
-    grid_var : np.ndarray
-        grid variable to be modified
-    original_fill : constant
-        original fill value used in (``grid_var``)
-    new_fill : constant
-        new fill value to be used in (``grid_var``)
-    new_dtype : np.dtype, optional
-        new data type to convert (``grid_var``) to
-
-    Returns
-    ----------
-    grid_var : xarray.Dataset
-        Input Dataset with correct fill value and dtype
+@njit(cache=True)
+def _small_angle_of_2_vectors(u, v):
     """
-
-    # locations of fill values
-    if original_fill is not None and np.isnan(original_fill):
-        fill_val_idx = np.isnan(grid_var)
-    else:
-        fill_val_idx = grid_var == original_fill
-
-    # convert to new data type
-    if new_dtype != grid_var.dtype and new_dtype is not None:
-        grid_var = grid_var.astype(new_dtype)
-
-    # ensure fill value can be represented with current integer data type
-    if np.issubdtype(new_dtype, np.integer):
-        int_min = np.iinfo(grid_var.dtype).min
-        int_max = np.iinfo(grid_var.dtype).max
-        # ensure new_fill is in range [int_min, int_max]
-        if new_fill < int_min or new_fill > int_max:
-            raise ValueError(
-                f"New fill value: {new_fill} not representable by"
-                f" integer dtype: {grid_var.dtype}"
-            )
-
-    # ensure non-nan fill value can be represented with current float data type
-    elif np.issubdtype(new_dtype, np.floating) and not np.isnan(new_fill):
-        float_min = np.finfo(grid_var.dtype).min
-        float_max = np.finfo(grid_var.dtype).max
-        # ensure new_fill is in range [float_min, float_max]
-        if new_fill < float_min or new_fill > float_max:
-            raise ValueError(
-                f"New fill value: {new_fill} not representable by"
-                f" float dtype: {grid_var.dtype}"
-            )
-    else:
-        raise ValueError(
-            f"Data type {grid_var.dtype} not supported" f"for grid variables"
-        )
-
-    # replace all zeros with a fill value
-    grid_var[fill_val_idx] = new_fill
-
-    return grid_var
-
-
-def _inv_jacobian(x0, x1, y0, y1, z0, z1, x_i_old, y_i_old):
-    """Calculate the inverse Jacobian matrix for a given set of parameters.
+    Compute the smallest angle between two vectors using the new _angle_of_2_vectors.
 
     Parameters
     ----------
-    x0 : float
-        Description of x0.
-    x1 : float
-        Description of x1.
-    y0 : float
-        Description of y0.
-    y1 : float
-        Description of y1.
-    z0 : float
-        Description of z0.
-    z1 : float
-        Description of z1.
-    x_i_old : float
-        Description of x_i_old.
-    y_i_old : float
-        Description of y_i_old.
+    u : numpy.ndarray
+        The first 3D vector.
+    v : numpy.ndarray
+        The second 3D vector.
 
     Returns
     -------
-    numpy.ndarray or None
-        The inverse Jacobian matrix if it is non-singular, or None if a singular matrix is encountered.
+    float
+        The smallest angle between `u` and `v` in radians.
+    """
+    v_norm_times_u = np.linalg.norm(v) * u
+    u_norm_times_v = np.linalg.norm(u) * v
+    vec_minus = v_norm_times_u - u_norm_times_v
+    vec_sum = v_norm_times_u + u_norm_times_v
+    angle_u_v_rad = 2 * np.arctan2(np.linalg.norm(vec_minus), np.linalg.norm(vec_sum))
+    return angle_u_v_rad
+
+
+@njit(cache=True)
+def _angle_of_2_vectors(u, v):
+    """
+    Calculate the angle between two 3D vectors `u` and `v` on the unit sphere in radians.
+
+    This function computes the angle between two vectors originating from the center of a unit sphere.
+    The result is returned in the range [0, 2π]. It can be used to calculate the span of a great circle arc (GCA).
+
+    Parameters
+    ----------
+    u : numpy.ndarray
+        The first 3D vector (float), originating from the center of the unit sphere.
+    v : numpy.ndarray
+        The second 3D vector (float), originating from the center of the unit sphere.
+
+    Returns
+    -------
+    float
+        The angle between `u` and `v` in radians, in the range [0, 2π].
 
     Notes
     -----
-    This function calculates the inverse Jacobian matrix based on the provided parameters. If the Jacobian matrix
-    is singular, a warning is printed, and None is returned.
+    - The direction of the angle (clockwise or counter-clockwise) is determined using the cross product of `u` and `v`.
+    - Special cases such as vectors aligned along the same longitude are handled explicitly.
     """
+    # Compute the cross product to determine the direction of the normal
+    normal = np.cross(u, v)
 
-    # d_dx = (x0 * x_i_old - x1 * x_i_old * z0 + y0 * y_i_old * z1 - y1 * y_i_old * z0 - y1 * y_i_old * z0)
-    # d_dy = 2 * (x0 * x_i_old * z1 - x1 * x_i_old * z0 + y0 * y_i_old * z1 - y1 * y_i_old * z0)
-    #
-    # # row 1
-    # J[0, 0] = y_i_old / d_dx
-    # J[0, 1] = (x0 * z1 - z0 * x1) / d_dy
-    # # row 2
-    # J[1, 0] = x_i_old / d_dx
-    # J[1, 1] = (y0 * z1 - z0 * y1) / d_dy
+    # Calculate the angle using arctangent of cross and dot products
+    angle_u_v_rad = np.arctan2(np.linalg.norm(normal), np.dot(u, v))
 
-    # The Jacobian Matrix
-    jacobian = [
-        [ac_utils._fmms(y0, z1, z0, y1), ac_utils._fmms(x0, z1, z0, x1)],
-        [2 * x_i_old, 2 * y_i_old],
-    ]
-    try:
-        inverse_jacobian = np.linalg.inv(jacobian)
-    except np.linalg.LinAlgError:
-        raise warnings("Warning: Singular Jacobian matrix encountered.")
-        return None
-
-    return inverse_jacobian
+    # Determine the direction of the angle
+    normal_z = np.dot(normal, np.array([0.0, 0.0, 1.0]))
+    if normal_z > 0:
+        # Counterclockwise direction
+        return angle_u_v_rad
+    elif normal_z == 0:
+        # Handle collinear vectors (same longitude)
+        if u[2] > v[2]:
+            return angle_u_v_rad
+        elif u[2] < v[2]:
+            return 2 * np.pi - angle_u_v_rad
+        else:
+            return 0.0  # u == v
+    else:
+        # Clockwise direction
+        return 2 * np.pi - angle_u_v_rad
 
 
-def _newton_raphson_solver_for_gca_constLat(
-    init_cart, gca_cart, max_iter=1000, verbose=False
-):
-    """Solve for the intersection point between a great circle arc and a
-    constant latitude.
+def _swap_first_fill_value_with_last(arr):
+    """Swap the first occurrence of INT_FILL_VALUE in each sub-array with the
+    last value in the sub-array.
 
-    Args:
-        init_cart (np.ndarray): Initial guess for the intersection point.
-        w0_cart (np.ndarray): First vector defining the great circle arc.
-        w1_cart (np.ndarray): Second vector defining the great circle arc.
-        max_iter (int, optional): Maximum number of iterations. Defaults to 1000.
-        verbose (bool, optional): Whether to print verbose output. Defaults to False.
+    Parameters:
+    ----------
+    arr (np.ndarray): A 3D numpy array where the swap will be performed.
 
     Returns:
-        np.ndarray or None: The intersection point or None if the solver fails to converge.
+    -------
+    np.ndarray: The modified array with the swaps made.
     """
-    tolerance = ERROR_TOLERANCE
-    w0_cart, w1_cart = gca_cart
-    error = float("inf")
-    constZ = init_cart[2]
-    y_guess = np.array(init_cart[0:2])
-    y_new = y_guess
+    # Find the indices of the first INT_FILL_VALUE in each sub-array
+    mask = arr == INT_FILL_VALUE
+    reshaped_mask = mask.reshape(arr.shape[0], -1)
+    first_true_indices = np.argmax(reshaped_mask, axis=1)
 
-    _iter = 0
+    # If no INT_FILL_VALUE is found in a row, argmax will return 0, we need to handle this case
+    first_true_indices[~np.any(reshaped_mask, axis=1)] = -1
 
-    while error > tolerance and _iter < max_iter:
-        f_vector = np.array(
-            [
-                np.dot(
-                    np.cross(w0_cart, w1_cart),
-                    np.array([y_guess[0], y_guess[1], constZ]),
-                ),
-                y_guess[0] * y_guess[0]
-                + y_guess[1] * y_guess[1]
-                + constZ * constZ
-                - 1.0,
-            ]
-        )
+    # Get the shape of the sub-arrays
+    subarray_shape = arr.shape[1:]
 
-        j_inv = _inv_jacobian(
-            w0_cart[0],
-            w1_cart[0],
-            w0_cart[1],
-            w1_cart[1],
-            w0_cart[2],
-            w1_cart[2],
-            y_guess[0],
-            y_guess[1],
-        )
+    # Calculate the 2D indices within each sub-array
+    valid_indices = first_true_indices != -1
+    first_true_positions = np.unravel_index(
+        first_true_indices[valid_indices], subarray_shape
+    )
 
-        if j_inv is None:
-            return None
+    # Create an index array for the last value in each sub-array
+    last_indices = np.full((arr.shape[0],), subarray_shape[0] * subarray_shape[1] - 1)
+    last_positions = np.unravel_index(last_indices, subarray_shape)
 
-        y_new = y_guess - np.matmul(j_inv, f_vector)
-        error = np.max(np.abs(y_guess - y_new))
-        y_guess = y_new
+    # Swap the first INT_FILL_VALUE with the last value in each sub-array
+    row_indices = np.arange(arr.shape[0])
 
-        if verbose:
-            print(f"Newton method iter: {_iter}, error: {error}")
-        _iter += 1
+    # Advanced indexing to swap values
+    (
+        arr[
+            row_indices[valid_indices], first_true_positions[0], first_true_positions[1]
+        ],
+        arr[
+            row_indices[valid_indices],
+            last_positions[0][valid_indices],
+            last_positions[1][valid_indices],
+        ],
+    ) = (
+        arr[
+            row_indices[valid_indices],
+            last_positions[0][valid_indices],
+            last_positions[1][valid_indices],
+        ],
+        arr[
+            row_indices[valid_indices], first_true_positions[0], first_true_positions[1]
+        ],
+    )
 
-    return np.append(y_new, constZ)
+    return arr
 
 
-# TODO: Consider re-implementation in the future / better integration with API
 def _get_cartesian_face_edge_nodes(
-    face_nodes_ind, face_edges_ind, edge_nodes_grid, node_x, node_y, node_z
+    face_node_conn, n_face, n_max_face_edges, node_x, node_y, node_z
 ):
     """Construct an array to hold the edge Cartesian coordinates connectivity
-    for a face in a grid.
+    for multiple faces in a grid.
 
     Parameters
     ----------
-    face_nodes_ind : np.ndarray, shape (n_nodes,)
-        The ith entry of Grid.face_node_connectivity, where n_nodes is the number of nodes in the face.
-    face_edges_ind : np.ndarray, shape (n_edges,)
-        The ith entry of Grid.face_edge_connectivity, where n_edges is the number of edges in the face.
-    edge_nodes_grid : np.ndarray, shape (n_edges, 2)
-        The entire Grid.edge_node_connectivity, where n_edges is the total number of edges in the grid.
-    node_x : np.ndarray, shape (n_nodes,)
-        The values of Grid.node_x, where n_nodes_total is the total number of nodes in the grid.
-    node_y : np.ndarray, shape (n_nodes,)
-        The values of Grid.node_y.
-    node_z : np.ndarray, shape (n_nodes,)
-        The values of Grid.node_z.
+    face_node_conn : np.ndarray
+        An array of shape (n_face, n_max_face_edges) containing the node indices for each face. Accessed through `grid.face_node_connectivity.value`.
+    n_face : int
+        The number of faces in the grid. Accessed through `grid.n_face`.
+    n_max_face_edges : int
+        The maximum number of edges for any face in the grid. Accessed through `grid.n_max_face_edges`.
+    node_x : np.ndarray
+        An array of shape (n_nodes,) containing the x-coordinate values of the nodes. Accessed through `grid.node_x`.
+    node_y : np.ndarray
+        An array of shape (n_nodes,) containing the y-coordinate values of the nodes. Accessed through `grid.node_y`.
+    node_z : np.ndarray
+        An array of shape (n_nodes,) containing the z-coordinate values of the nodes. Accessed through `grid.node_z`.
 
     Returns
     -------
-    face_edges : np.ndarray, shape (n_edges, 2, 3)
-        Face edge connectivity in Cartesian coordinates, where n_edges is the number of edges for the specific face.
+    face_edges_cartesian : np.ndarray
+        An array of shape (n_face, n_max_face_edges, 2, 3) containing the Cartesian coordinates of the edges
+        for each face. It might contain dummy values if the grid has holes.
 
-    Notes
-    -----
-    - The function assumes that the inputs are well-formed and correspond to the same face.
-    - The output array contains the Cartesian coordinates for each edge of the face.
+    Examples
+    --------
+    >>> face_node_conn = np.array(
+    ...     [
+    ...         [0, 1, 2, 3, 4],
+    ...         [0, 1, 3, 4, INT_FILL_VALUE],
+    ...         [0, 1, 3, INT_FILL_VALUE, INT_FILL_VALUE],
+    ...     ]
+    ... )
+    >>> n_face = 3
+    >>> n_max_face_edges = 5
+    >>> node_x = np.array([0, 1, 1, 0, 1, 0])
+    >>> node_y = np.array([0, 0, 1, 1, 2, 2])
+    >>> node_z = np.array([0, 0, 0, 0, 1, 1])
+    >>> _get_cartesian_face_edge_nodes(
+    ...     face_node_conn, n_face, n_max_face_edges, node_x, node_y, node_z
+    ... )
+    array([[[[    0,     0,     0],
+         [    1,     0,     0]],
+
+        [[    1,     0,     0],
+         [    1,     1,     0]],
+
+        [[    1,     1,     0],
+         [    0,     1,     0]],
+
+        [[    0,     1,     0],
+         [    1,     2,     1]],
+
+        [[    1,     2,     1],
+         [    0,     0,     0]]],
+
+
+       [[[    0,     0,     0],
+         [    1,     0,     0]],
+
+        [[    1,     0,     0],
+         [    0,     1,     0]],
+
+        [[    0,     1,     0],
+         [    1,     2,     1]],
+
+        [[    1,     2,     1],
+         [    0,     0,     0]],
+
+        [[INT_FILL_VALUE, INT_FILL_VALUE, INT_FILL_VALUE],
+        [INT_FILL_VALUE, INT_FILL_VALUE, INT_FILL_VALUE]]],
+
+
+       [[[    0,     0,     0],
+         [    1,     0,     0]],
+
+        [[    1,     0,     0],
+         [    0,     1,     0]],
+
+        [[    0,     1,     0],
+         [    0,     0,     0]],
+
+        [[INT_FILL_VALUE, INT_FILL_VALUE, INT_FILL_VALUE],
+         [INT_FILL_VALUE, INT_FILL_VALUE, INT_FILL_VALUE]],
+
+        [[INT_FILL_VALUE, INT_FILL_VALUE, INT_FILL_VALUE],
+         [INT_FILL_VALUE, INT_FILL_VALUE, INT_FILL_VALUE]]]])
     """
+    # Shift node connections to create edge connections
+    face_node_conn_shift = np.roll(face_node_conn, -1, axis=1)
 
-    # Create a mask that is True for all values not equal to INT_FILL_VALUE
-    mask = face_edges_ind != INT_FILL_VALUE
+    # Construct edge connections by combining original and shifted node connections
+    face_edge_conn = np.array([face_node_conn, face_node_conn_shift]).T.swapaxes(0, 1)
 
-    # Use the mask to select only the elements not equal to INT_FILL_VALUE
-    valid_edges = face_edges_ind[mask]
-    face_edges = edge_nodes_grid[valid_edges]
+    # swap the first occurrence of INT_FILL_VALUE with the last value in each sub-array
+    face_edge_conn = _swap_first_fill_value_with_last(face_edge_conn)
 
-    # Ensure counter-clockwise order of edge nodes
-    # Start with the first two nodes
-    face_edges[0] = [face_nodes_ind[0], face_nodes_ind[1]]
+    # Get the indices of the nodes from face_edge_conn
+    face_edge_conn_flat = face_edge_conn.reshape(-1)
 
-    for idx in range(1, len(face_edges)):
-        if face_edges[idx][0] != face_edges[idx - 1][1]:
-            # Swap the node index in this edge if not in counter-clockwise order
-            face_edges[idx] = face_edges[idx][::-1]
+    valid_mask = face_edge_conn_flat != INT_FILL_VALUE
 
-    # Fetch coordinates for each node in the face edges
-    cartesian_coordinates = np.array(
-        [
-            [[node_x[node], node_y[node], node_z[node]] for node in edge]
-            for edge in face_edges
-        ]
+    # Get the valid node indices
+    valid_edges = face_edge_conn_flat[valid_mask]
+
+    #  Create an array to hold the Cartesian coordinates of the edges
+    face_edges_cartesian = np.full(
+        (len(face_edge_conn_flat), 3), INT_FILL_VALUE, dtype=float
     )
 
-    return cartesian_coordinates
+    # Fill the array with the Cartesian coordinates of the edges
+    face_edges_cartesian[valid_mask, 0] = node_x[valid_edges]
+    face_edges_cartesian[valid_mask, 1] = node_y[valid_edges]
+    face_edges_cartesian[valid_mask, 2] = node_z[valid_edges]
+
+    return face_edges_cartesian.reshape(n_face, n_max_face_edges, 2, 3)
 
 
 def _get_lonlat_rad_face_edge_nodes(
-    face_nodes_ind, face_edges_ind, edge_nodes_grid, node_lon, node_lat
+    face_node_conn, n_face, n_max_face_edges, node_lon, node_lat
 ):
-    """Construct an array to hold the edge lat lon in radian connectivity for a
-    face in a grid.
+    """Construct an array to hold the edge latitude and longitude in radians
+    connectivity for multiple faces in a grid.
 
     Parameters
     ----------
-    face_nodes_ind : np.ndarray, shape (n_nodes,)
-        The ith entry of Grid.face_node_connectivity, where n_nodes is the number of nodes in the face.
-    face_edges_ind : np.ndarray, shape (n_edges,)
-        The ith entry of Grid.face_edge_connectivity, where n_edges is the number of edges in the face.
-    edge_nodes_grid : np.ndarray, shape (n_edges, 2)
-        The entire Grid.edge_node_connectivity, where n_edges is the total number of edges in the grid.
-    node_lon : np.ndarray, shape (n_nodes,)
-        The values of Grid.node_lon.
-    node_lat : np.ndarray, shape (n_nodes,)
-        The values of Grid.node_lat, where n_nodes_total is the total number of nodes in the grid.
+    face_node_conn : np.ndarray
+        An array of shape (n_face, n_max_face_edges) containing the node indices for each face. Accessed through `grid.face_node_connectivity.value`.
+    n_face : int
+        The number of faces in the grid. Accessed through `grid.n_face`.
+    n_max_face_edges : int
+        The maximum number of edges for any face in the grid. Accessed through `grid.n_max_face_edges`.
+    node_lon : np.ndarray
+        An array of shape (n_nodes,) containing the longitude values of the nodes in degrees. Accessed through `grid.node_lon`.
+    node_lat : np.ndarray
+        An array of shape (n_nodes,) containing the latitude values of the nodes in degrees. Accessed through `grid.node_lat`.
+
     Returns
     -------
-    face_edges : np.ndarray, shape (n_edges, 2, 3)
-        Face edge connectivity in Cartesian coordinates, where n_edges is the number of edges for the specific face.
+    face_edges_lonlat_rad : np.ndarray
+        An array of shape (n_face, n_max_face_edges, 2, 2) containing the latitude and longitude coordinates
+        in radians for the edges of each face. It might contain dummy values if the grid has holes.
 
     Notes
     -----
-    - The function assumes that the inputs are well-formed and correspond to the same face.
-    - The output array contains the Latitude and longitude coordinates in radian for each edge of the face.
+    If the grid has holes, the function will return an entry of dummy value faces_edges_coordinates[i] filled with INT_FILL_VALUE.
     """
 
-    # Create a mask that is True for all values not equal to INT_FILL_VALUE
-    mask = face_edges_ind != INT_FILL_VALUE
+    # Convert node coordinates to radians
+    node_lon_rad = np.deg2rad(node_lon)
+    node_lat_rad = np.deg2rad(node_lat)
 
-    # Use the mask to select only the elements not equal to INT_FILL_VALUE
-    valid_edges = face_edges_ind[mask]
-    face_edges = edge_nodes_grid[valid_edges]
+    # Shift node connections to create edge connections
+    face_node_conn_shift = np.roll(face_node_conn, -1, axis=1)
 
-    # Ensure counter-clockwise order of edge nodes
-    # Start with the first two nodes
-    face_edges[0] = [face_nodes_ind[0], face_nodes_ind[1]]
+    # Construct edge connections by combining original and shifted node connections
+    face_edge_conn = np.array([face_node_conn, face_node_conn_shift]).T.swapaxes(0, 1)
 
-    for idx in range(1, len(face_edges)):
-        if face_edges[idx][0] != face_edges[idx - 1][1]:
-            # Swap the node index in this edge if not in counter-clockwise order
-            face_edges[idx] = face_edges[idx][::-1]
+    # swap the first occurrence of INT_FILL_VALUE with the last value in each sub-array
+    face_edge_conn = _swap_first_fill_value_with_last(face_edge_conn)
 
-    # Fetch coordinates for each node in the face edges
-    lonlat_coordinates = np.array(
-        [
-            [
-                [
-                    np.mod(np.deg2rad(node_lon[node]), 2 * np.pi),
-                    np.deg2rad(node_lat[node]),
-                ]
-                for node in edge
-            ]
-            for edge in face_edges
-        ]
+    # Get the indices of the nodes from face_edge_conn
+    face_edge_conn_flat = face_edge_conn.reshape(-1)
+
+    valid_mask = face_edge_conn_flat != INT_FILL_VALUE
+
+    # Get the valid node indices
+    valid_edges = face_edge_conn_flat[valid_mask]
+
+    # Create an array to hold the latitude and longitude in radians for the edges
+    face_edges_lonlat_rad = np.full(
+        (len(face_edge_conn_flat), 2), INT_FILL_VALUE, dtype=float
     )
 
-    return lonlat_coordinates
+    # Fill the array with the latitude and longitude in radians for the edges
+    face_edges_lonlat_rad[valid_mask, 0] = node_lon_rad[valid_edges]
+    face_edges_lonlat_rad[valid_mask, 1] = node_lat_rad[valid_edges]
+
+    return face_edges_lonlat_rad.reshape(n_face, n_max_face_edges, 2, 2)
