@@ -45,11 +45,13 @@ def _read_topology(
             else:
                 conn_arr = kwargs[conn]
 
-            ds[conn] = xr.DataArray(
-                data=_process_connectivity(conn_arr, fill_value, start_index),
+            conn_da = xr.DataArray(
+                data=conn_arr,
                 dims=ugrid.CONNECTIVITY[conn]["dims"],
                 attrs=ugrid.CONNECTIVITY[conn]["attrs"],
             )
+
+            ds[conn] = _process_connectivity(conn_da, fill_value, start_index)
 
     return ds
 
@@ -58,9 +60,17 @@ def _process_connectivity(conn, orig_fv, start_index):
     """Internal helper for processing connectivity variables, standardizing
     fill values and converting to zero-index."""
 
+    if orig_fv is None:
+        # no fill values, adjust start index
+        return conn - start_index
+
     if orig_fv != INT_FILL_VALUE:
         conn = _replace_fill_values(conn, orig_fv, INT_FILL_VALUE, INT_DTYPE)
 
-    conn[conn != INT_FILL_VALUE] -= start_index
+    conn = xr.where(
+        conn != INT_FILL_VALUE,
+        conn - start_index,
+        conn,
+    )
 
     return conn
