@@ -7,7 +7,6 @@ class HoloviewsBackend:
     visualization."""
 
     def __init__(self):
-        self.backend = None
         self.matplotlib_backend = mpl.get_backend()
 
     def assign(self, backend: str):
@@ -19,15 +18,13 @@ class HoloviewsBackend:
             Plotting backend to use, one of 'matplotlib', 'bokeh'
         """
 
-        if backend not in ["bokeh", "matplotlib"]:
+        if backend not in ["bokeh", "matplotlib", None]:
             raise ValueError(
                 f"Unsupported backend. Expected one of ['bokeh', 'matplotlib'], but received {backend}"
             )
-
-        if backend != self.backend:
+        if backend is not None and backend != hv.Store.current_backend:
             # only call hv.extension if it needs to be changed
             hv.extension(backend)
-            self.backend = backend
 
     def reset_mpl_backend(self):
         """Resets the default backend for the ``matplotlib`` module."""
