@@ -1,5 +1,5 @@
 from __future__ import annotations
-from typing import TYPE_CHECKING, Union
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from uxarray.core.dataset import UxDataset
@@ -12,16 +12,20 @@ import uxarray.core.dataset
 from uxarray.grid import Grid
 import warnings
 
+from copy import deepcopy
+
+from uxarray.remap.utils import _remap_grid_parse
+
 
 def _inverse_distance_weighted_remap(
-    source_grid,
-    destination_grid,
-    source_data,
-    remap_to="nodes",
-    coord_type="spherical",
+    source_grid: Grid,
+    destination_grid: Grid,
+    source_data: np.ndarray,
+    remap_to: str = "face centers",
+    coord_type: str = "spherical",
     power=2,
     k=8,
-):
+) -> np.ndarray:
     """Inverse Distance Weighted Remapping between two grids.
 
     Parameters:
@@ -49,7 +53,7 @@ def _inverse_distance_weighted_remap(
     """
 
     if power > 5:
-        warnings.warn("It is recommended not to exceed a power of 5.0.")
+        warnings.warn("It is recommended not to exceed a power of 5.0.", UserWarning)
     if k > source_grid.n_node:
         raise ValueError(
             f"Number of nearest neighbors to be used in the calculation is {k}, but should not exceed the "
@@ -60,88 +64,18 @@ def _inverse_distance_weighted_remap(
             f"Number of nearest neighbors to be used in the calculation is {k}, but should be greater than 1"
         )
 
+    # ensure array is a np.ndarray
     source_data = np.asarray(source_data)
-    n_elements = source_data.shape[-1]
 
-    if n_elements == source_grid.n_node:
-        source_data_mapping = "nodes"
-    elif n_elements == source_grid.n_face:
-        source_data_mapping = "face centers"
-    elif n_elements == source_grid.n_edge:
-        source_data_mapping = "edge centers"
-    else:
-        raise ValueError(
-            f"Invalid source_data shape. The final dimension should match the number of corner "
-            f"nodes ({source_grid.n_node}), edge nodes ({source_grid.n_edge}), or face centers ({source_grid.n_face}) "
-            f"in the source grid, but received: {source_data.shape}"
-        )
-
-    if coord_type == "spherical":
-        if remap_to == "nodes":
-            lon, lat = (
-                destination_grid.node_lon.values,
-                destination_grid.node_lat.values,
-            )
-        elif remap_to == "face centers":
-            lon, lat = (
-                destination_grid.face_lon.values,
-                destination_grid.face_lat.values,
-            )
-        elif remap_to == "edge centers":
-            lon, lat = (
-                destination_grid.edge_lon.values,
-                destination_grid.edge_lat.values,
-            )
-        else:
-            raise ValueError(
-                f"Invalid remap_to. Expected 'nodes', 'edge centers', or 'face centers', "
-                f"but received: {remap_to}"
-            )
-
-        _source_tree = source_grid.get_ball_tree(coordinates=source_data_mapping)
-
-        dest_coords = np.vstack([lon, lat]).T
-
-        distances, nearest_neighbor_indices = _source_tree.query(dest_coords, k=k)
-
-    elif coord_type == "cartesian":
-        if remap_to == "nodes":
-            x, y, z = (
-                destination_grid.node_x.values,
-                destination_grid.node_y.values,
-                destination_grid.node_z.values,
-            )
-        elif remap_to == "face centers":
-            x, y, z = (
-                destination_grid.face_x.values,
-                destination_grid.face_y.values,
-                destination_grid.face_z.values,
-            )
-        elif remap_to == "edge centers":
-            x, y, z = (
-                destination_grid.edge_x.values,
-                destination_grid.edge_y.values,
-                destination_grid.edge_z.values,
-            )
-        else:
-            raise ValueError(
-                f"Invalid remap_to. Expected 'nodes', 'edge centers', or 'face centers', "
-                f"but received: {remap_to}"
-            )
-
-        _source_tree = source_grid.get_kd_tree(coordinates=source_data_mapping)
-
-        dest_coords = np.vstack([x, y, z]).T
-
-        distances, nearest_neighbor_indices = _source_tree.query(dest_coords, k=k)
-
-    else:
-        raise ValueError(
-            f"Invalid coord_type. Expected either 'spherical' or 'cartesian', but received {coord_type}"
-        )
-
-    if nearest_neighbor_indices.ndim > 1:
-        nearest_neighbor_indices = nearest_neighbor_indices.squeeze()
+    _, distances, nearest_neighbor_indices = _remap_grid_parse(
+        source_data,
+        source_grid,
+        destination_grid,
+        coord_type,
+        remap_to,
+        k=k,
+        query=True,
+    )
 
     weights = 1 / (distances**power + 1e-6)
     weights /= np.sum(weights, axis=1, keepdims=True)
@@ -155,8 +89,8 @@ def _inverse_distance_weighted_remap(
 
 def _inverse_distance_weighted_remap_uxda(
     source_uxda: UxDataArray,
-    destination_obj: Union[Grid, UxDataArray, UxDataset],
-    remap_to: str = "nodes",
+    destination_grid: Grid,
+    remap_to: str = "face centers",
     coord_type: str = "spherical",
     power=2,
     k=8,
@@ -167,8 +101,8 @@ def _inverse_distance_weighted_remap_uxda(
     ---------
     source_uxda : UxDataArray
         Source UxDataArray for remapping
-    destination_obj : Grid, UxDataArray, UxDataset
-        Destination for remapping
+    destination_grid : Grid
+        Destination grid for remapping
     remap_to : str, default="nodes"
         Location of where to map data, either "nodes", "edge centers", or "face centers"
     coord_type : str, default="spherical"
@@ -202,16 +136,6 @@ def _inverse_distance_weighted_remap_uxda(
     destination_dims = list(source_uxda.dims)
     destination_dims[-1] = destination_dim
 
-    if isinstance(destination_obj, Grid):
-        destination_grid = destination_obj
-    elif isinstance(
-        destination_obj,
-        (uxarray.core.dataarray.UxDataArray, uxarray.core.dataset.UxDataset),
-    ):
-        destination_grid = destination_obj.uxgrid
-    else:
-        raise ValueError("TODO: Invalid Input")
-
     # perform remapping
     destination_data = _inverse_distance_weighted_remap(
         source_uxda.uxgrid,
@@ -222,33 +146,29 @@ def _inverse_distance_weighted_remap_uxda(
         power,
         k,
     )
+
+    # preserve only non-spatial coordinates
+    destination_coords = deepcopy(source_uxda.coords)
+    if destination_dim in destination_coords:
+        del destination_coords[destination_dim]
+
     # construct data array for remapping variable
     uxda_remap = uxarray.core.dataarray.UxDataArray(
         data=destination_data,
         name=source_uxda.name,
         dims=destination_dims,
         uxgrid=destination_grid,
+        coords=destination_coords,
     )
-    # add remapped variable to existing UxDataset
-    if isinstance(destination_obj, uxarray.core.dataset.UxDataset):
-        destination_obj[source_uxda.name] = uxda_remap
-        return destination_obj
-
-    # construct a UxDataset from remapped variable and existing variable
-    elif isinstance(destination_obj, uxarray.core.dataset.UxDataArray):
-        uxds = destination_obj.to_dataset()
-        uxds[source_uxda.name] = uxda_remap
-        return uxds
 
     # return UxDataArray with remapped variable
-    else:
-        return uxda_remap
+    return uxda_remap
 
 
 def _inverse_distance_weighted_remap_uxds(
     source_uxds: UxDataset,
-    destination_obj: Union[Grid, UxDataArray, UxDataset],
-    remap_to: str = "nodes",
+    destination_grid: Grid,
+    remap_to: str = "face centers",
     coord_type: str = "spherical",
     power=2,
     k=8,
@@ -259,7 +179,7 @@ def _inverse_distance_weighted_remap_uxds(
     ---------
     source_uxds : UxDataset
         Source UxDataset for remapping
-    destination_obj : Grid, UxDataArray, UxDataset
+    destination_grid : Grid
         Destination for remapping
     remap_to : str, default="nodes"
         Location of where to map data, either "nodes", "edge centers", or "face centers"
@@ -271,20 +191,11 @@ def _inverse_distance_weighted_remap_uxds(
     k : int, default=8
         Number of nearest neighbors to consider in the weighted calculation.
     """
-
-    if isinstance(destination_obj, Grid):
-        destination_uxds = uxarray.core.dataset.UxDataset(uxgrid=destination_obj)
-    elif isinstance(destination_obj, uxarray.core.dataset.UxDataArray):
-        destination_uxds = destination_obj.to_dataset()
-    elif isinstance(destination_obj, uxarray.core.dataset.UxDataset):
-        destination_uxds = destination_obj
-    else:
-        raise ValueError
-
+    destination_uxds = uxarray.UxDataset(uxgrid=destination_grid)
     for var_name in source_uxds.data_vars:
-        destination_uxds = _inverse_distance_weighted_remap_uxda(
+        destination_uxds[var_name] = _inverse_distance_weighted_remap_uxda(
             source_uxds[var_name],
-            destination_uxds,
+            destination_grid,
             remap_to,
             coord_type,
             power,
