@@ -1,6 +1,86 @@
 import numpy as np
 import sys
 
+from numba import njit
+
+
+@njit(cache=True)
+def clip(a, a_min, a_max):
+    return np.clip(a, a_min, a_max)
+
+
+@njit(cache=True)
+def arcsin(x):
+    return np.arcsin(x)
+
+
+@njit(cache=True)
+def all(a):
+    """Numba decorated implementation of ``np.all()``
+
+    See Also
+    --------
+    numpy.all
+    """
+
+    return np.all(a)
+
+
+@njit(cache=True)
+def isclose(a, b, rtol=1e-05, atol=1e-08):
+    """Numba decorated implementation of ``np.isclose()``
+
+    See Also
+    --------
+    numpy.isclose
+    """
+
+    return np.isclose(a, b, rtol=rtol, atol=atol)
+
+
+@njit(cache=True)
+def allclose(a, b, rtol=1e-05, atol=1e-08):
+    """Numba decorated implementation of ``np.allclose()``
+
+    See Also
+    --------
+    numpy.allclose
+    """
+    return np.allclose(a, b, rtol=rtol, atol=atol)
+
+
+@njit(cache=True)
+def cross(a, b):
+    """Numba decorated implementation of ``np.cross()``
+
+    See Also
+    --------
+    numpy.cross
+    """
+    return np.cross(a, b)
+
+
+@njit(cache=True)
+def dot(a, b):
+    """Numba decorated implementation of ``np.dot()``
+
+    See Also
+    --------
+    numpy.dot
+    """
+    return np.dot(a, b)
+
+
+@njit(cache=True)
+def norm(x):
+    """Numba decorated implementation of ``np.linalg.norm()``
+
+    See Also
+    --------
+    numpy.linalg.norm
+    """
+    return np.linalg.norm(x)
+
 
 def _fmms(a, b, c, d):
     """
@@ -22,7 +102,7 @@ def _fmms(a, b, c, d):
 
     Example
     -------
-    >>> _fmms(3.0,2.0,1.0,1.0)
+    >>> _fmms(3.0, 2.0, 1.0, 1.0)
     5.0
 
     Reference
@@ -116,8 +196,8 @@ def dot_fma(v1, v2):
 
 
 def _two_prod_fma(a, b):
-    """
-    Error-free transformation of the product of two floating-point numbers using FMA, such that a * b = x + y exactly.
+    """Error-free transformation of the product of two floating-point numbers
+    using FMA, such that a * b = x + y exactly.
 
     Parameters
     ----------
@@ -147,9 +227,9 @@ def _two_prod_fma(a, b):
 
 
 def _err_fmac(a, b, c):
-    """
-    Error-free transformation for the FMA operation. such that x = FMA(a,b,c) and a * b + c = x + y + z exactly.
-    Thhis function is only available in round to the nearest mode and takes approximately 17 flops
+    """Error-free transformation for the FMA operation. such that x =
+    FMA(a,b,c) and a * b + c = x + y + z exactly. Thhis function is only
+    available in round to the nearest mode and takes approximately 17 flops.
 
     Parameters
     ----------
@@ -187,8 +267,8 @@ def _err_fmac(a, b, c):
 
 
 def _two_sum(a, b):
-    """
-    Error-free transformation of the sum of two floating-point numbers such that a + b = x + y exactly
+    """Error-free transformation of the sum of two floating-point numbers such
+    that a + b = x + y exactly.
 
     Parameters
     ----------
@@ -216,8 +296,8 @@ def _two_sum(a, b):
 
 
 def _fast_two_mult(a, b):
-    """
-    Error-free transformation of the product of two floating-point numbers such that a * b = x + y exactly.
+    """Error-free transformation of the product of two floating-point numbers
+    such that a * b = x + y exactly.
 
     This function is faster than the _two_prod_fma function.
 
@@ -236,7 +316,6 @@ def _fast_two_mult(a, b):
     Vincent Lefèvre, Nicolas Louvet, Jean-Michel Muller, Joris Picot, and Laurence Rideau. 2023.
     Accurate Calculation of Euclidean Norms Using Double-word Arithmetic.
     ACM Trans. Math. Softw. 49, 1, Article 1 (March 2023), 34 pages. https://doi.org/10.1145/3568672
-
     """
     import pyfma
 
